% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{aberrantSep}
\alias{aberrantSep}
\title{Assesses whether two reads in a pair are aberrantly separated with respect to the reference genome}
\usage{
aberrantSep(gpairs, distance = 10000)
}
\arguments{
\item{gpairs}{a \code{GAlignmentPairs} object}

\item{distance}{the minimum distance in base pairs between two reads in 
order to call them aberrantly separated}
}
\value{
logical vector of the same length as \code{gpairs}
}
\description{
Determines whether separation between first and last read in a
GAlignmentPairs is greater than some distance.
}
\details{
Evaluates to TRUE if
}
