\name{academic_articles}
\docType{data}
\alias{academic_articles}
\title{Sample Wikipedia articles containing the term ``academic''}
\description{
	This data set contains titles and file contents for 80 sample
	documents scraped from Wikipedia. All documents contain the
	term ``academic.''
}
\usage{academic_articles}
\format{
	List with two entries.
	Each entry is a character vector corresponding to 80 articles
}
\source{Scraped from 80 Wikipedia articles}
\keyword{datasets}