% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf2}
\alias{acf2}
\title{Compute the autocorrelation at a specific lag}
\usage{
acf2(x, lag.max = 10, type = c("correlation", "covariance", "partial"),
  plot = FALSE, na.action = na.omit, demean = TRUE, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{lag.max}{integer specifying the autocorrelation lag}

\item{type}{character string}

\item{plot}{logical}

\item{na.action}{function to handle missing observations}

\item{demean}{logical}

\item{...}{additional arguments passed to \code{acf}}
}
\description{
A wrapper for \code{acf} that returns a single autocorrelation for
a specified lag.
}
\seealso{
\code{\link{acf}}
}
