% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.info.R
\name{add.info}
\alias{add.info}
\title{Add seqinfo to bed files (for cram file version)}
\usage{
add.info(data)
}
\arguments{
\item{data}{}
}
\value{
data
}
\description{
Add seqinfo to bed files (for cram file version)
}
