% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{addFocalDupsFlankingAmplicon}
\alias{addFocalDupsFlankingAmplicon}
\title{Add focal amplicons that flank an amplicon seed}
\usage{
addFocalDupsFlankingAmplicon(object, rp, params)
}
\arguments{
\item{object}{a \code{AmpliconGraph}}

\item{rp}{a \code{GAlignmentPairs} object}

\item{params}{a list of parameters for amplicon analyses}
}
\description{
Amplicons selected to seed a graph have a fold change of at least log2(2.75),
or nearly 3-fold the diploid genome. Lower-copy amplicons flanking the seed
amplicons are added to the graph if they have a fold change of at least
log2(1.75), or nearly 2-fold the diploid genome. To assess whether a low-copy
amplicon is 'flanking' an amplicon seed, all read pairs that aligned to the
amplicon seeds (see \code{get_readpairs}) are passed to this function in
object \code{rp}. The read pairs are converted to segments by
\code{readPairsAsSegments} and an assessment of whether flanking duplications
are linked is given by \code{linkedDuplicatedRanges}. In particular, this
function checks whether there are at least \code{minimum_count} fragments
that connect the amplicon seeds to the flanking regions. Finally, the flanks
are added to the graph by the \code{addFlanks} function.
}
\details{
REFACTORING: overly complicated with many abstractions.  Needs an
example to step through
}
\examples{
\dontrun{
  path <- file.path(system.file(package="svcnvs"), "tests/testthat")
  ag <- readRDS(file.path(path, "addFocalDups.ffab104.rds"))
  params <- ampliconParams()
  ag <- addFocalDupsFlankingAmplicon(ag, rp, params)
}
}
\seealso{
\code{\link{ampliconParams}}
}
