% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{improperAlignmentFlags}
\alias{improperAlignmentFlags}
\alias{properAlignmentFlags}
\alias{improperAlignmentParams}
\alias{properAlignmentParams}
\title{Creates a default set of flags for reading improperly paired alignments}
\usage{
improperAlignmentFlags()

properAlignmentFlags()

improperAlignmentParams(
  flag = improperAlignmentFlags(),
  what = c("flag", "mrnm", "mpos", "mapq", "qname"),
  ...
)

properAlignmentParams(
  flag = properAlignmentFlags(),
  what = c("flag", "mrnm", "mpos", "mapq"),
  ...
)
}
\arguments{
\item{flag}{A length-two integer vector as provided by \code{improperAlignmentFlags}}

\item{what}{A character vector (see \code{ScanBamParam} for details)}

\item{...}{additional arguments to \code{ScanBamParam} such as \code{mapqFilter}}
}
\description{
These functions are wrappers for \code{scanBamFlag} and
\code{ScanBamParam} in the \code{Rsamtools} package.
}
\examples{
require(Rsamtools)

flags <- scanBamFlag(isDuplicate=FALSE,
                     isPaired=TRUE,
                     isUnmappedQuery=FALSE,
                     hasUnmappedMate=FALSE,
                     isProperPair=FALSE)
flags2 <- improperAlignmentFlags()
identical(flags, flags2) #TRUE
print(flags)
 
flags <- scanBamFlag(isDuplicate=FALSE,
                     isPaired=TRUE,
                     isUnmappedQuery=FALSE,
                     hasUnmappedMate=FALSE,
                     isProperPair=TRUE)
flags2 <- properAlignmentFlags()
identical(flags, flags2) #TRUE
print(flags)

flags <- improperAlignmentFlags()
print(flags)
params <- ScanBamParam(flag = flags, what=c("flag", "mrnm", "mpos", "mapq"))
params2 <- improperAlignmentParams()
print(params2)
identical(params, params2) #TRUE

flags <- properAlignmentFlags()
print(flags)
params <- ScanBamParam(flag = flags, what=c("flag", "mrnm", "mpos", "mapq"))
params2 <- properAlignmentParams()
identical(params, params2) #TRUE
print(params2)
}
\seealso{
See \code{\link[Rsamtools]{scanBamFlag}} for complete details
  and \code{\link{improperAlignmentParams}} for a wrapper of this
  function that generates a \code{ScanBamParam} object using these
  flags.

See \code{\link[Rsamtools]{ScanBamParam}} and
  \code{\link[Rsamtools]{bamFlag}} in
  \code{Rsamtools} for full documentation.  See \code{improperAlignmentFlags} for the
  default set of flags.
}
