% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentViews-class.R
\docType{class}
\name{AlignmentViews2-class}
\alias{AlignmentViews2}
\alias{AlignmentViews2-class}
\alias{[,AlignmentViews2,missing,character,ANY-method}
\alias{[,AlignmentViews2,missing,character-method}
\alias{[,AlignmentViews2,missing,logical,ANY-method}
\alias{[,AlignmentViews2,missing,logical-method}
\alias{[,AlignmentViews2,missing,numeric,ANY-method}
\alias{[,AlignmentViews2,missing,numeric-method}
\alias{improperPaths}
\alias{improperPaths,AlignmentViews2-method}
\title{Container for storing views to extracted alignments}
\usage{
AlignmentViews2(bviews = BamViews(), path)

improperPaths(object)

\S4method{improperPaths}{AlignmentViews2}(object)

\S4method{[}{AlignmentViews2,missing,numeric,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{AlignmentViews2,missing,logical,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{AlignmentViews2,missing,character,ANY}(x, i, j, ...,
  drop = FALSE)
}
\arguments{
\item{bviews}{A \code{BamViews} object}

\item{path}{file path for storing alignment data}

\item{object}{a \code{AlignmentViews2} object}

\item{x}{a \code{AlignmentViews2} object}

\item{i}{index for \code{rowRanges} elements}

\item{j}{index for samples}

\item{...}{ignored}

\item{drop}{ignored}
}
\value{
a character-vector of file paths
}
\description{
Improper read pairs are extracted from BAM files to augment CNV-
and rearrangement-analyses

AlignmentViews2 objects inherit from \code{BamViews}.  In addition
to containing file paths to one or more bam files, AlignmentViews2
objects contain file paths to a subset of improperly paired reads
that were / will be parsed from the bams.  There will be one parsed
file per bam file.

Accessor for file names containing improper read pair alignments
}
\section{Slots}{

\describe{
\item{\code{improperAlignmentPaths}}{character-vector of file paths}
}}
\examples{
  library(Rsamtools)
  require(TestBams)
  extdata <- system.file("extdata", package="TestBams")
  bam.file <- list.files(extdata, pattern="\\\\.bam$", full.name=TRUE)
  bv <- BamViews(bam.file)
  fname <- tempfile()
  aviews <- AlignmentViews2(bv, fname)
  validObject(aviews[, 1])
  aviews
  improperPaths(aviews)
}

