% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Amplicon-class.R
\docType{class}
\name{amplicons}
\alias{amplicons}
\alias{amplicons<-}
\alias{ampliconRanges}
\alias{ampliconRanges<-}
\alias{assemblyGaps}
\alias{driver}
\alias{graph}
\alias{graph<-}
\alias{isAmplicon}
\alias{isSomatic}
\alias{nonampliconRanges}
\alias{queryRanges}
\alias{queryRanges<-}
\alias{ampliconNames}
\alias{AmpliconGraph-class}
\alias{addNode,character,AmpliconGraph,ANY-method}
\alias{addNode,character,AmpliconGraph-method}
\alias{amplicons,AmpliconGraph-method}
\alias{amplicons<-,AmpliconGraph-method}
\alias{amplicons<-,AmpliconGraph,ANY-method}
\alias{assemblyGaps,AmpliconGraph-method}
\alias{ampliconRanges<-,AmpliconGraph-method}
\alias{assemblyGaps<-,AmpliconGraph,ANY-method}
\alias{ampliconRanges,AmpliconGraph-method}
\alias{ampliconRanges,GRanges-method}
\alias{graph,AmpliconGraph-method}
\alias{graph,AmpliconGraph-class}
\alias{driver,AmpliconGraph-method}
\alias{edges,AmpliconGraph,ANY-method}
\alias{edges,AmpliconGraph-method}
\alias{graph<-,AmpliconGraph-method}
\alias{graph,AmpliconGraph,ANY-method}
\alias{isAmplicon,AmpliconGraph-method}
\alias{isAmplicon,GRanges-method}
\alias{isSomatic,GRanges-method}
\alias{length,AmpliconGraph-method}
\alias{nodes,AmpliconGraph-method}
\alias{names<-,AmpliconGraph,ANY-method}
\alias{nonampliconRanges,AmpliconGraph-method}
\alias{nonAmpliconRanges,AmpliconGraph-method}
\alias{nonampliconRanges,GRanges-method}
\alias{nonAmpliconGranges,GRanges-method}
\alias{numEdges,AmpliconGraph-method}
\alias{numNodes,AmpliconGraph-method}
\alias{queryRanges,AmpliconGraph-method}
\alias{queryRanges<-,AmpliconGraph-method}
\alias{queryRanges,AmpliconGraph,ANY-method}
\alias{ranges,AmpliconGraph-method}
\alias{ranges<-,AmpliconGraph-method}
\alias{ranges,AmpliconGraph,ANY-method}
\title{A class for representing somatic amplicons}
\usage{
amplicons(object)

amplicons(object) <- value

ampliconRanges(object)

ampliconRanges(object) <- value

assemblyGaps(object)

driver(object)

graph(object)

graph(object) <- value

isAmplicon(object)

isSomatic(object)

nonampliconRanges(object)

queryRanges(object)

queryRanges(object) <- value

ampliconNames(amplicons)

\S4method{addNode}{character,AmpliconGraph,ANY}(node, object, edges)

\S4method{amplicons}{AmpliconGraph}(object)

\S4method{amplicons}{AmpliconGraph}(object) <- value

\S4method{assemblyGaps}{AmpliconGraph}(object)

\S4method{ampliconRanges}{AmpliconGraph}(object) <- value

\S4method{ampliconRanges}{AmpliconGraph}(object)

\S4method{ampliconRanges}{GRanges}(object)

\S4method{graph}{AmpliconGraph}(object)

\S4method{driver}{AmpliconGraph}(object)

\S4method{edges}{AmpliconGraph,ANY}(object, which, ...)

\S4method{graph}{AmpliconGraph}(object) <- value

\S4method{isAmplicon}{AmpliconGraph}(object)

\S4method{isAmplicon}{GRanges}(object)

\S4method{isSomatic}{GRanges}(object)

\S4method{length}{AmpliconGraph}(x)

\S4method{nodes}{AmpliconGraph}(object)

\S4method{names}{AmpliconGraph,ANY}(x) <- value

\S4method{nonampliconRanges}{AmpliconGraph}(object)

\S4method{nonampliconRanges}{GRanges}(object)

\S4method{numEdges}{AmpliconGraph}(object)

\S4method{numNodes}{AmpliconGraph}(object)

\S4method{queryRanges}{AmpliconGraph}(object)

\S4method{queryRanges}{AmpliconGraph}(object) <- value

\S4method{ranges}{AmpliconGraph}(x, use.names = TRUE, use.mcols = FALSE, ...)

\S4method{ranges}{AmpliconGraph}(x) <- value
}
\arguments{
\item{object}{a \code{AmpliconGraph} object}

\item{value}{genomic amplicon intervals as a \code{GRanges} object}

\item{amplicons}{genomic intervals of amplicons as a \code{GRanges} object}

\item{node}{name of amplicon}

\item{edges}{name of edge (e.g., concatenate two amplicon names)}

\item{which}{ignored}

\item{...}{ignored}

\item{x}{a \code{AmpliconGraph}}

\item{use.names}{logical}

\item{use.mcols}{logical}
}
\description{
Somatic amplicons are often grouped by a bridge-breakage-fusion
cycles.  We represent the set of amplicons identified for a sample
as a graph with amplicons as the nodes. An edge drawn between two
nodes indicates that five or more read pairs with aberrant spacing
support the link between these amplicons.  As the class is an
extension of the \code{graphNEL} class, plot methods to visualize
the graph are readily available.
}
\seealso{
See \code{sv_amplicon_exp} for generating a list of
  \code{AmpliconGraphs} for an experiment.
}
