% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{AmpliconGraph}
\alias{AmpliconGraph}
\title{Constructor for AmpliconGraph}
\usage{
AmpliconGraph(
  ranges = GRanges(),
  filters,
  params = ampliconParams(border_size = 150000, overhang = 5000)
)
}
\arguments{
\item{ranges}{a \code{GRanges} object of the segmented normalized copy number}

\item{filters}{a list of germline CNV and sequence-based filters}

\item{params}{a list of parameters}
}
\description{
Constructor for \code{AmpliconGraph}.
}
