% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{ampliconParams}
\alias{ampliconParams}
\title{Lists germline filters and parameters for filtering amplicons}
\usage{
ampliconParams(AMP_THR = log2(2.75), LOW_THR = log2(1.75),
  border_size = 10000, overhang = 25000, MIN_WIDTH = 2000,
  MIN_SEGMEAN_DIFF = 0.05, min.gapwidth = 3000, maxgap = 5e+05,
  frequency = 5, minimum_maxdist = 50, minimum_count = 5,
  bad_bins = GRanges())
}
\arguments{
\item{AMP_THR}{numeric threshold for high copy amplicon}

\item{LOW_THR}{numeric a lower threshold used when considering
amplicons that are bridged by improper read pairs to a high copy
amplicon}

\item{border_size}{used to construct a query (GRanges object) for
additional amplicons neighboring a focal amplicon. TODO: more
detail needed.}

\item{overhang}{An integer indicating how much to expand the
germline filter on each size.  Using \code{overlapsAny}, a
determination is made whether an amplicon is part-germline (any
overlap) or no-germline (no overlap).  If the amplicon is
completely within the extended germline filter, the amplicon is
considered fully-germline.}

\item{MIN_WIDTH}{length-one integer vector indicating the minimum size of amplicon (see also \code{joinNearGRanges})}

\item{MIN_SEGMEAN_DIFF}{length-one numeric vector.  Adjacent segments whose means differ by less than this value are candidates for merging by \code{{joinNearGRanges}}}

\item{min.gapwidth}{length-one numeric vector passed to the \code{reduce} method that merges adacent segments with comparable segment means}

\item{maxgap}{length-one numeric vector. Passed to \code{findOverlaps} when
evaluating whether the amplicon ranges overlap with other amplicons (see
\code{linkNearAmplicons}).}

\item{frequency}{minimum number of read pairs}

\item{minimum_maxdist}{TODO}

\item{minimum_count}{a length-one integer vector: minimum number of
improperly paired reads required to link two amplicons}

\item{bad_bins}{a \code{GRanges} object of problematic bins}
}
\value{
Returns a list of germline filters and some of the hard
  thresholds used in the amplicon analysis.
}
\description{
Lists germline filters and parameters for filtering amplicons
}
\examples{
params <- ampliconParams("hg19")

}
