% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amplicons-data}
\alias{amplicons-data}
\title{An example AmpliconGraph}
\format{
a \code{AmpliconGraph} object.
}
\usage{
data(amplicons)
}
\description{
An \code{AmpliconGraph} for sample CGOV2T.
}
\examples{
data(amplicons)
ampliconRanges(amplicons)
}
\keyword{datasets}
