% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{annotateBlatRecords}
\alias{annotateBlatRecords}
\title{Annotate blat records}
\usage{
annotateBlatRecords(blat, tag.sequences)
}
\arguments{
\item{blat}{\code{data.frame} of blat records from \code{readBlat}}

\item{tag.sequences}{\code{data.frame} of read sequences}
}
\description{
Annotate blat records
}
\examples{
data(tags)
extdata <- system.file("extdata", package="trellis")
blat.file <- file.path(extdata, "blat_alignment.txt")
blat_aln <- readBlat(blat.file)
records <- annotateBlatRecords(blat_aln, tags)
annotateBlatRecords(records, tags)

}
