% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.R
\name{annotateRecurrent}
\alias{annotateRecurrent}
\title{Annotate table of amplified genes with linked drivers and the number of
samples for which a driver is linked}
\usage{
annotateRecurrent(rcnvs, grl)
}
\arguments{
\item{rcnvs}{a table of recurrent CNVs}

\item{grl}{a \code{GRangesList} of the segmented copy number (each element is a sample)}
}
\description{
Annotate table of amplified genes with linked drivers and the number of
samples for which a driver is linked
}
\seealso{
See \code{\link{recurrentDrivers}} for making a table of recurrent amplicons
}
