% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess-class.R
\name{assays,PreprocessViews2-method}
\alias{assays,PreprocessViews2-method}
\title{Accessor for reading assay data saved to disk}
\usage{
\S4method{assays}{PreprocessViews2}(x, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{a \code{PreprocessViews2} object}

\item{withDimnames}{ignored}

\item{...}{ignored}
}
\value{
a R x C matrix, where R is the length of
  \code{rowRanges(x)} and C is the number of samples given by
  \code{ncol(x)}
}
\description{
We have adapted the assays method to read assay data from disk
using a \code{PreprocessViews2} object.
}
\details{
REFACTOR: both ... and withDimnames are ignored.  we should just
provide a different generic/method.
}
