% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{binCounts}
\alias{binCounts}
\title{Counts reads mapped to genomic intervals}
\usage{
binCounts(object, scan_param)
}
\arguments{
\item{object}{A \code{BamViews} object}

\item{scan_param}{An optional \code{ScanBamParam} object specifying any
parameters to consider when importing reads from bam files included in the 
\code{BamViews} object.  If omitted, a \code{scanBamParams} object 
will automatically be created to include all ranges from 
the \code{BamViews} object and to ignore unmapped and duplicate reads.}
}
\value{
An integer-vector of counts corresponding to each interval in the \code{BamViews} object.
}
\description{
Utilizes the \code{countBam()} function in the \code{Rsamtools} package to 
count number of reads mapped to ranges determined by 
a \code{ScanBamParams()} object.
}
\examples{
library(Rsamtools)
library(svbams)
library(svfilters.hg19)
data(bins1kb)
extdir <- system.file("extdata", package="svbams", mustWork=TRUE)
bamfile <- file.path(extdir, "cgov10t.bam")
bins <- keepSeqlevels(bins1kb, "chr3", pruning.mode = "coarse")
bins <- subsetByOverlaps(bins, GRanges("chr3", IRanges(59600000, 61000000)))
bviews <- BamViews(bamRanges=bins, bamPaths=bamfile)
counts <- binCounts(bviews)
counts
length(bamRanges(bviews)) == length(counts) #TRUE

}
