% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{binFragments}
\alias{binFragments}
\title{Counts the number of read fragments mapped to bins}
\usage{
binFragments(galp, bins)
}
\arguments{
\item{galp}{a \code{GAlignmentPairs} object}

\item{bins}{a \code{GRanges} object}
}
\value{
an integer-vector of counts corresponding to each range in the \code{GRanges} object
}
\description{
For each bin provided by a \code{GRanges object}, \code{binFragments}
counts the number of fragments, determined by read pairs in a
\code{GAlignmentPair} object, that overlap that bin
}
\examples{
\dontrun{
library(Rsamtools)
library(svbams)
library(GenomicAlignments)
bam_path <- system.file("extdata", package="svbams", mustWork=TRUE)
bamfile <- file.path(bam_path, "cgov10t.bam")
data(bins, package='svbams')
param <- ScanBamParam(flag=scanBamFlag(isDuplicate=FALSE,
                                       isSecondaryAlignment=FALSE)) 
galp <- readGAlignmentPairs(bamfile, param=param)
fragmentCounts <- binFragments(galp, bins)
}
}
