% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{binGCCorrect}
\alias{binGCCorrect}
\title{Correct coverage for GC content}
\usage{
binGCCorrect(bins)
}
\arguments{
\item{bins}{a \code{GRanges} object. This must contain a metadata column of numerics 
named \code{std_cnt} that is generally generated by \code{binNormalize()}.}
}
\value{
A numeric vector of normalized counts for each bin.
}
\description{
Normalizes the coverage for a set of bins such that there is no correlation between GC content and coverage.
}
\details{
The relationship of counts to GC content is determined by fitting a loess curve. 
The lowess curve is subtracted from the counts for each bin to remove any GC bias.
}
\examples{
library(Rsamtools)
library(svbams)
library(svfilters.hg19)
library(trellis)
data(bins1kb)
extdir <- system.file("extdata", package="svbams", mustWork=TRUE)
bamfile <- file.path(extdir, "cgov10t.bam")
bins <- keepSeqlevels(bins1kb, "chr3", pruning.mode = "coarse")
bins <- subsetByOverlaps(bins, GRanges("chr3", IRanges(59600000, 61000000)))
bviews <- BamViews(bamRanges=bins, bamPaths=bamfile)
bins$cnt <- trellis::binnedCounts(bviews)
std_cnt <- binNormalize(bins)
bins$std_cnt <- std_cnt
head(trellis::binGCCorrect(bins))
}
