% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{binNormalize}
\alias{binNormalize}
\title{Normalize counts}
\usage{
binNormalize(bins, centerby = c("mode", "median"))
}
\arguments{
\item{bins}{a \code{GRanges} object}

\item{centerby}{mode or median}
}
\value{
A numeric vector of normalized counts.
}
\description{
Log-transforms the counts for each bin and centers them by by the 
mode (default) or median of the counts for the autosomes.
}
\details{
The \code{bins} argument to \code{binNormalize()} must be a \code{GRanges} object with  
chromosome names in the format [chr1, chr2, ...] and not [1, 2, ...] or any other naming convention.  
There must be a metadata column of integers named \code{cnt} corresponding to the raw counts 
for each bin.    

The output vector from \code{binNormalize()} is generally appended back to \code{bins}
as a metadata column named \code{std_cnt} for GC normalization by \code{binGCCorrect()}.
}
\examples{
library(Rsamtools)
library(svbams)
library(svfilters.hg19)
data(bins1kb)
extdir <- system.file("extdata", package="svbams", mustWork=TRUE)
bamfile <- file.path(extdir, "cgov10t.bam")
bins <- keepSeqlevels(bins1kb, "chr3", pruning.mode = "coarse")
bins <- subsetByOverlaps(bins, GRanges("chr3", IRanges(59600000, 61000000)))
bviews <- BamViews(bamRanges=bins, bamPaths=bamfile)
bins$cnt <- binnedCounts(bviews)
head(binNormalize(bins))
std_cnt <- binNormalize(bins)
bins$std_cnt <- std_cnt
}
