% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blat_unmapped}
\alias{blat_unmapped}
\title{BLAT alignment of reads that were unmapped by ELAND}
\format{a \code{data.frame} of blat records}
\description{
This is a subset of reads that were involved in an unmapped-mapped
pair.  The mapped reads were near a rearrangement supported by
improperly paired mapped-mapped reads ('mapped-mapped' indicates
both reads were alignmed by ELAND).  We used BLAT to assess whether
a split read alignment for the unmapped read supports the new
sequence junction.
}
