% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{blatScores}
\alias{blatScores}
\title{blatScores assesses whether the improper read pairs at a
rearrangement junction provide strong support of the rearrangement}
\usage{
blatScores(blat, tags, id, thr = 0.8)
}
\arguments{
\item{blat}{a \code{data.frame} of results from  command-line blat}

\item{tags}{a \code{data.frame} containing read names and the original alignment locations}

\item{id}{a character-vector of sample identifiers}

\item{thr}{a length-one numeric vector indicating the fraction of
reads at a rearrangement that must pass the read-level QC.}
}
\description{
In the following, we refer to a 'record' as one row in the table of
blat output -- i.e., one (of possibly many) alignments for a read.
This function adds an indicator for whether the blat alignment is
consistent with the original alignment (\code{is_overlap}), an
indicator of whether it passes quality control (\code{passQC}) (see
details), the id of the rearrangement (\code{rearrangement}), and
the sample id (\code{id}).
}
\details{
\strong{Read-level QC:} 

For each record, we evaluate

\enumerate{

\item whether the matching score is at least 90

\item whether the size of the target alignment (Tstart - Tend) is
  less than 120bp and more than 80bp

\item whether the blat alignment overlaps with any of the original
  alignment

}

The records are then grouped by Qname.  For each Qname, we compute
  the number of reads with a score above 90 and within the
  specified size range.  In addition, we compute the number of
  reads with a score above 90, within the specified size range, and
  that overlap with the original alignment.  The Qname passes QC if
  both sums evaluate to 1.  That is, a read passes QC only if there
  is a single record with a high BLAT score within the specified
  size range and this single record overlaps with the original
  alignment.

\strong{Rearrangement-level QC:}

Given a pass / fail designation for each read by the above
  analysis, we group the reads by the rearrangement id.  A
  rearrangement passes QC if > 80% of the reads supporting the
  rearrangement pass QC.
}
\examples{
 qnames <- c(paste0(letters[1:10], "_R1"),
             paste0(letters[1:10], "_R2"))
 ## only 1 location
 numberAlignedLocations <- rep(1, length(qnames))
 matchScores <- rep(95, length(qnames))
 Tend <-  150
 Tstart <- 51
 ## Made up output from blat
 sblat <- data.frame(Qname=qnames,
                     match=matchScores,
                     Tend=Tend,
                     Tstart=Tstart,
                     Tname=rep("chr1", length(qnames)),
                     strand=rep("+", length(qnames)),
                     stringsAsFactors=FALSE)

## Made up information on a rearrangement, including the locations
##  at which the reads were originally aligned

 stags <- data.frame(qname=rep(letters[1:10], 2),
                     read=rep(c("R1", "R2"), each=10),
                     seqnames=rep("chr1", 10),
                     strand=rep("+", 10),
                     start=Tstart,
                     end=Tend,
                     seq=replicate(10, paste(sample(c("g","c"), 10, replace=TRUE),
                         collapse="")),
                     stringsAsFactors=FALSE)
 stags$id <- "CGOV32T"
 stags$rearrangement.id <- "1-3"
 ## For each tag, calculate the number of near-perfect matches of the
 ## right size that overlap with eland.  If the number is 0 or more
 ## than 1, then the tag 'fails'.
 blat <- trellis:::annotateBlatRecords(sblat, stags)
 nrow(blat) == 20
 s <- blatScores(sblat, stags, "SOME_ID")
 head(s)
}
