% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R
\docType{methods}
\name{calls}
\alias{calls}
\alias{calls,StructuralVariant,ANY-method}
\alias{calls,StructuralVariant-method}
\alias{calls<-}
\alias{calls<-,StructuralVariant-method}
\title{Type of deletion}
\usage{
calls(object)

calls(object) <- value

\S4method{calls}{StructuralVariant}(object)

\S4method{calls}{StructuralVariant}(object) <- value
}
\arguments{
\item{object}{a \code{StructuralVariant}}

\item{value}{a character string indicating the type of deletion}
}
\description{
An accessor for the deletion calls.  The deletion calls are
hemizygous, hemizygous+, homozygous, and homozygous+.  The '+'
denotes that improper read pairs support the deletion identified
from the read depth analysis.
}

