% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAcopy-class.R
\name{cbs2granges}
\alias{cbs2granges}
\title{Coerce a DNAcopy object to GRanges}
\usage{
cbs2granges(x, seq_info)
}
\arguments{
\item{x}{a \code{DNAcopy} object}

\item{seq_info}{a \code{Seqinfo} object}
}
\value{
a \code{GRanges} object
}
\description{
Coerce a DNAcopy object to GRanges
}
