% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqinfo-utils.R
\docType{methods}
\name{autosomeNames}
\alias{autosomeNames}
\alias{autosomeNames,BamViews-method}
\alias{autosomeNames,GRanges-method}
\alias{chromosome}
\alias{chromosome,GAlignments-method}
\alias{chromosome,GRanges-method}
\alias{chromosome,RangedSummarizedExperiment-method}
\alias{chromosome,StructuralVariant-method}
\title{Extract autosomal (human) sequence levels}
\usage{
autosomeNames(object)

chromosome(object)

\S4method{chromosome}{GRanges}(object)

\S4method{chromosome}{RangedSummarizedExperiment}(object)

\S4method{chromosome}{GAlignments}(object)

\S4method{chromosome}{StructuralVariant}(object)

\S4method{autosomeNames}{BamViews}(object)

\S4method{autosomeNames}{GRanges}(object)
}
\arguments{
\item{object}{a \code{GRanges} object}
}
\description{
Extract autosomal (human) sequence levels

Return seqnames of a \code{GRanges} object as a character vector
}

