% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Transcripts-class.R
\docType{class}
\name{ClippedTranscripts}
\alias{ClippedTranscripts}
\alias{ClippedTranscripts-class}
\alias{ClippedTranscripts,missing,missing-method}
\alias{left,ClippedTranscripts-method}
\alias{right,ClippedTranscripts-method}
\title{A container for clipped transcripts}
\usage{
ClippedTranscripts(
  transcripts,
  i,
  left = GRangesList(),
  right = GRangesList(),
  names = c("5prime", "3prime")
)

\S4method{ClippedTranscripts}{missing,missing}(
  transcripts,
  i,
  left = GRangesList(),
  right = GRangesList(),
  names = c("5prime", "3prime")
)

\S4method{left}{ClippedTranscripts}(x, ...)

\S4method{right}{ClippedTranscripts}(x, ...)
}
\arguments{
\item{transcripts}{a \code{Transcripts} object}

\item{i}{an integer vector}

\item{left}{a \code{GRangesList} for the transcripts to the left of the junction}

\item{right}{a \code{GRangesList} for the transcripts to the right of the junction}

\item{names}{character vector indicating strand of transcripts ('5prime' or '3prime').}

\item{x}{a \code{ClippedTranscripts} object}

\item{...}{ignored}
}
\description{
A container for clipped transcripts
}
\section{Slots}{

\describe{
\item{\code{left}}{a \code{GRangesList}}

\item{\code{right}}{a \code{GRangesList}}

\item{\code{names}}{a character vector of transcript names}
}}

\examples{
ClippedTranscripts()
}
