% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{codingJunctions}
\alias{codingJunctions}
\title{Select sequence junctions in which both the five-prime and three-prime genomic regions occur within a transcript}
\usage{
codingJunctions(jxns, build)
}
\arguments{
\item{jxns}{a \code{GRanges} representation of the sequence junctions}

\item{build}{a character-string}
}
\description{
For the 3-prime genomic region, the sequence junction must occur within a transcript (usually intronic).  For the 5-prime sequence junction, the genomic region can either be within 5-kb upstream of the transcript (defined as the promoter) or within the transcript.
}
\examples{
# TODO
}
