% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{countParam}
\alias{countParam}
\title{Generate parameters that specify which reads in a bam file to import}
\usage{
countParam(
  gr,
  flags = scanBamFlag(isUnmappedQuery = FALSE, isDuplicate = FALSE)
)
}
\arguments{
\item{gr}{A \code{GRanges} object}

\item{flags}{An integer vector as returned by \code{Rsamtools::scanBamFlag()}.
Unmapped and duplicate reads are omitted by default.}
}
\value{
A \code{\link[Rsamtools]{ScanBamParam}} object
}
\description{
A wrapper function for \code{Rsamtools::ScanBamParam()}.
}
\examples{
library(Rsamtools)
library(svfilters.hg19)
data(bins1kb)
scan_param <- countParam(bins1kb)
scan_param
class(scan_param) #ScanBamParam
}
