% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{countReads2}
\alias{countReads2}
\title{Count reads in bamRanges of a BamViews object}
\usage{
countReads2(object, scan_param)
}
\arguments{
\item{object}{A \code{BamViews} object}

\item{scan_param}{a \code{ScanBamParams} object}
}
\description{
This function counts the number of reads mapped to each interval
given by \code{bamRanges(object)}.
}
