% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPaths-class.R
\docType{class}
\name{DataPaths-class}
\alias{DataPaths-class}
\alias{DataPaths}
\alias{[[,DataPaths,character,ANY-method}
\alias{[,DataPaths,character,ANY,ANY-method}
\title{A class for storing file paths to intermediate data types}
\usage{
DataPaths(path = tempdir(), rootname = "sv_data", dryrun = FALSE)

\S4method{[[}{DataPaths,character,ANY}(x, i)

\S4method{[}{DataPaths,character,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{path}{A length-one character vector specifying where the top-level directory should reside}

\item{rootname}{A length-one character vector providing the name of the directory to create under \code{path}}

\item{dryrun}{logical.  If \code{FALSE}, all directories are
created. Otherwise, only the character-vector is returned.}

\item{x}{a \code{DataPaths} object}

\item{i}{a length-one character vector}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\description{
A class for storing file paths to intermediate data types

Intermediate data is saved to disk to avoid rerunning long-steps in
computation. The Paths constructor initializes a vector of file
paths for storing different types of intermediate data.
}
\examples{
## by default, file directories are not created
DataPaths()
## Create directories for intermediate data types
all_paths <- DataPaths(dryrun=FALSE)
all_paths
## Just list the 'data' subdirectories
topics(all_paths)

## Subsetting by character-string uses grep
all_paths[["count"]]
all_paths["count"]
all_paths["data"] ## many paths have 'data' as part of string
all_paths["gc"]

}
\seealso{
\code{\link{topics}}
}
