% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{deletion_call}
\alias{deletion_call}
\title{Creates a StructuralVariant object, linking genomic intervals
(deletions) to read pairs supporting the rearrangement}
\usage{
deletion_call(preprocess, gr_filters, param = DeletionParam())
}
\arguments{
\item{preprocess}{A \code{list} object generated by \code{preprocessData}}

\item{gr_filters}{A \code{GRanges} object of germline CNVs and
sequence-based filters identified by low mappability and/or GC
content}

\item{param}{A \code{DeletionParam} object}
}
\description{
Identifies somatic, focal hemizygous and homozygous deletions, then
links these deletions with properly and improperly paired reads.
}
\seealso{
See \link{germlineFilters} for identifiying somatic deletions.
}
