% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeletionParam-class.R
\docType{class}
\name{DeletionParam-class}
\alias{DeletionParam-class}
\alias{DeletionParam}
\alias{bamSeqLevelsStyle}
\alias{homozygousThr}
\alias{hemizygousThr}
\alias{maximumWidth}
\alias{maximumProportionInFilter}
\alias{minSegmeanDiff}
\alias{maxRatioObservedToExpected}
\alias{minimumWidth}
\alias{minimumGapWidth}
\alias{minFlankingHemizygous}
\alias{minFlankingHomozygous}
\alias{minRPsHemizygous}
\alias{maxRPsHomozygous}
\title{Parameter class for calling deletions}
\usage{
DeletionParam(
  min_width = 2000,
  max_width = 2e+06,
  min_gapwidth = 10000,
  homozygous_thr = -3,
  hemizygous_thr = -0.7,
  tumor_purity = 1,
  min_RPs_in_hemizygous = 1,
  max_RPs_in_homozygous = 100,
  nflanking_hemizygous = 5L,
  nflanking_homozygous = 5L,
  maxRatioObservedToExpected = 2L,
  max_proportion_in_filter = 0.75,
  min_segmean_diff = -0.5,
  bam_seqlevels_style = "UCSC",
  remove_hemizygous = TRUE
)

bamSeqLevelsStyle(object)

homozygousThr(object)

hemizygousThr(object)

maximumWidth(object)

maximumProportionInFilter(object)

minSegmeanDiff(object)

maxRatioObservedToExpected(object)

minimumWidth(object)

minimumGapWidth(object)

minFlankingHemizygous(object)

minFlankingHomozygous(object)

minRPsHemizygous(object)

maxRPsHomozygous(object)
}
\arguments{
\item{min_width}{length-one numeric vector indicating the minimum width of a
CNV segment.}

\item{max_width}{length-one numeric vector indicating the maximum width of a
CNV segment.}

\item{min_gapwidth}{length-one numeric vector used to reduce deletion calls.
Provided that the segment means of adjacent deletions are qualitatively
similar, two deletions less than min_gapwidth apart will be merged into a
single CNV.}

\item{homozygous_thr}{length-one numeric vector indicating the maximum log2
ratio expected for a homozygous deletion}

\item{hemizygous_thr}{length-one numeric vector indicating the maximum log2
ratio expected for a hemizygous deletion}

\item{tumor_purity}{length-one numeric vector in [0, 1]}

\item{min_RPs_in_hemizygous}{length-one numeric vector indicating the minimum
number of rearranged read pairs that can be spanned by a hemizygous
deletion. [PARAMETER NOT USED]}

\item{max_RPs_in_homozygous}{length-one numeric vector indicating the maximum
number of read pairs allowed to be spanned by a homozygous deletion.
[PARAMETER NOT USED]}

\item{nflanking_hemizygous}{length-one numeric vector indicating the minimum
number of rearranged RPs required for calling a hemizygous deletion with
higher confidence. Hemizygous deletions with rearranged RPs are given
higher priority over hemizygous deletions supported only by the tag
density.}

\item{nflanking_homozygous}{length-one numeric vector indicating the minimum
number of rearranged RPs required for calling a homozygous deletion with
higher confidence. Homozygous deletions with rearranged RPs are given
higher priority over homozygous deletions supported only by the tag
density.}

\item{maxRatioObservedToExpected}{length-one numeric vector [IGNORED?]}

\item{max_proportion_in_filter}{length-one numeric vector indicating the
maximum proportion of a CNV that can be covered by one of the filters. The
filters are germline CNV, low mappability, low GC / unassembled regions,
and outliers observed in a bin for 2 or more normal samples.}

\item{min_segmean_diff}{length one numeric vector indicating how much
less the segment mean for a deletion must be relative to the median
log ratio of bins encompassed in an interval surrounding the
segment of length equal to 15 times the width of the deleted segment.}

\item{bam_seqlevels_style}{character string indicating style of seqnames}

\item{object}{a \code{DeletionParam}}

\item{build}{length-1 character vector indicating genome build (must be hg18
or hg19)}
}
\value{
a \code{DeletionParam} object
}
\description{
Parameter class for calling deletions
}
\section{Slots}{

\describe{
\item{\code{homozygous_thr}}{length-one numeric vector
indicating the maximum fold change expected for a homozygous
deletion}

\item{\code{hemizygous_thr}}{length-one numeric vector
indicating the maximum fold change expected for a hemizygous
deletion}

\item{\code{min_width}}{length-one numeric vector indicating the minimum
width of a CNV segment.}

\item{\code{max_width}}{length-one numeric vector indicating the maximum
width of a CNV segment.}

\item{\code{min_gapwidth}}{length-one numeric vector used to reduce
deletion calls. Provided that the segment means of adjacent
deletions are qualitatively similar, two deletions less than
min_gapwidth apart will be merged into a single CNV.}

\item{\code{min_RPs_in_hemizygous}}{length-one numeric vector indicating
the minimum number of rearranged read pairs that can be spanned
by a hemizygous deletion. [PARAMETER NOT USED]}

\item{\code{max_RPs_in_homozygous}}{length-one numeric vector indicating
the maximum number of read pairs allowed to be spanned by a
homozygous deletion.  [PARAMETER NOT USED]}

\item{\code{nflanking_hemizygous}}{length-one numeric vector indicating the
minimum number of rearranged RPs required for calling a
hemizygous deletion with higher confidence.  Hemizygous deletions
with rearranged RPs are given higher priority over hemizygous
deletions supported only by the tag density.}

\item{\code{nflanking_homozygous}}{length-one numeric vector indicating the
minimum number of rearranged RPs required for calling a
homozygous deletion with higher confidence.  Homozygous deletions
with rearranged RPs are given higher priority over homozygous
deletions supported only by the tag density.}

\item{\code{maxRatioObservedToExpected}}{length-one numeric vector [IGNORED?]}

\item{\code{max_proportion_in_filter}}{length-one numeric vector indicating
the maximum proportion of a CNV that can be covered by one of the
filters.  The filters are germline CNV, low mappability, low GC /
unassembled regions, and outliers observed in a bin for 2 or more
normal samples.}

\item{\code{min_segmean_diff}}{length one numeric vector indicating how much
less the segment mean for a deletion must be relative to the median
log ratio of bins encompassed in an interval surrounding the
segment of length equal to 15 times the width of the deleted segment.}

\item{\code{bam_seqlevels_style}}{character string indicating style of seqnames}

\item{\code{tumor_purity}}{estimated fraction of sequenced DNA that is tumor-derived}

\item{\code{remove_hemizygous}}{whether to remove "hemizygous" calls ('hemizygous+' calls are not removed).}
}}

\examples{
DeletionParam()
}
\keyword{internal}
