% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deletions}
\alias{deletions}
\title{A deletions object of class StructuralVariant}
\format{a \code{StructuralVariant} object containing several
  deletions.}
\usage{
data(deletions)
}
\description{
A \code{StructuralVariant} object containing data supporting
deletions.
}
\examples{
data(deletions)
variant(deletions)
improper(deletions)
deletions[1]
improper(deletions[1])
calls(deletions)
}
\keyword{datasets}
