% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPaths-class.R
\name{dirCreate}
\alias{dirCreate}
\title{Recursively create directories from supplied file paths}
\usage{
dirCreate(x)
}
\value{
character-vector of file paths
}
\description{
Recursively create directories from supplied file paths
}
\seealso{
\code{\link{dir.create}}
}
\keyword{internal}

