% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAcopy-class.R
\docType{methods}
\name{DNAcopy-class}
\alias{DNAcopy-class}
\alias{DNAcopy}
\alias{as}
\title{Class for storing circular binary segmentation results}
\arguments{
\item{from}{S3 class DNAcopy}
}
\value{
a \code{GRanges} object
}
\description{
Class for storing circular binary segmentation results

Coerce \code{DNAcopy} object to \code{GRanges}
}
\examples{
library(DNAcopy)
## example from DNAcopy
genomdat <- rnorm(500, sd=0.1) +
              rep(c(-0.2,0.1,1,-0.5,0.2,-0.5,0.1,-0.2),
                  c(137,87,17,49,29,52,87,42))
chrom <- rep(1:2,c(290,210))
maploc <- c(1:290,1:210)
test1 <- segment(CNA(genomdat, chrom, maploc))
as(test1, "GRanges")
}
\seealso{
\code{\link[DNAcopy]{segment}}
}
