% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{dt2ga2}
\alias{dt2ga2}
\title{Convert dataframe to GAlignment (for CRAM)}
\usage{
dt2ga2(dt)
}
\arguments{
\item{dt}{}
}
\value{
GAlignment object that is ready to be paired with GAlignmentPairs
}
\description{
Convert dataframe to GAlignment (for CRAM)
}
