% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{duplicatedGAlignmentPairs}
\alias{duplicatedGAlignmentPairs}
\title{Determines which rows from a \code{GAlignmentPairs} object are duplicates of a row with a smaller subscript}
\usage{
duplicatedGAlignmentPairs(galp)
}
\arguments{
\item{galp}{a \code{GAlignmentPairs} object}
}
\value{
a logical vector corresponding to each row in the \code{GAlignmentPairs} object
}
\description{
Given a \code{GAlignmentPairs} object, \code{duplicatedGAlignmentPairs} returns a logical vector with one element for each row.
}
\examples{
library(Rsamtools)
library(svbams)
library(GenomicAlignments)
bam_path <- system.file("extdata", package="svbams", mustWork=TRUE)
bamfile <- file.path(bam_path, "cgov10t.bam")
data(bins, package='svbams')
param1 <- ScanBamParam(flag=scanBamFlag(isDuplicate=FALSE,
                                        isSecondaryAlignment=FALSE),
                       which=bins)
param1b <- ScanBamParam(flag=scanBamFlag(isDuplicate=FALSE,
                                        isSecondaryAlignment=FALSE))
galp1 <- readGAlignmentPairs(bamfile, param=param1)
galp1 <- galp1[do.call(order, as.data.frame(galp1))]
galp1b <- readGAlignmentPairs(bamfile, param=param1b)
galp1b <- subsetByOverlaps(galp1b, bins, ignore.strand=TRUE)
galp1b <- galp1b[do.call(order, as.data.frame(galp1b))]
identical(galp1, galp1b) ## FALSE because galp1 has duplicates
galp1 <- galp1[!duplicatedGAlignmentPairs(galp1)]
identical(galp1, galp1b)

}
