% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExonSubset-class.R
\docType{class}
\name{inRearrangement.left}
\alias{ExonSubset}
\alias{ExonSubset-class}
\alias{ExonSubset-method}
\alias{inRearrangement.left}
\alias{inRearrangement.left,ExonSubset-method}
\alias{inRearrangement.right}
\alias{inRearrangement.right,ExonSubset-method}
\alias{name.left}
\alias{name.left,ExonSubset-method}
\alias{name.right}
\alias{name.right,}
\alias{name.right,ExonSubset-method}
\alias{names,ExonSubset-method}
\title{Container for storing information to subset transcripts}
\usage{
inRearrangement.left(object)

inRearrangement.right(object)

name.left(object)

name.right(object)

ExonSubset(name.left = character(), inRearrangement.left = LogicalList(),
  name.right = character(), inRearrangement.right = LogicalList())

\S4method{name.left}{ExonSubset}(object)

\S4method{name.right}{ExonSubset}(object)

\S4method{names}{ExonSubset}(x)

\S4method{inRearrangement.left}{ExonSubset}(object)

\S4method{inRearrangement.right}{ExonSubset}(object)
}
\arguments{
\item{object}{a \code{ExonSubset} object}

\item{name.left}{name of left transcript}

\item{inRearrangement.left}{\code{LogicalList} subsetting the left transcripts}

\item{name.right}{name of right transcript}

\item{inRearrangement.right}{\code{LogicalList} subsetting the right transcripts}

\item{x}{an \code{ExonSubset} object}
}
\value{
a \code{LogicalList}

a \code{LogicalList}
}
\description{
Container for storing information to subset transcripts
}
\section{Slots}{

\describe{
\item{\code{name.left}}{character string}

\item{\code{inRearrangement.left}}{a \code{LogicalList}}

\item{\code{name.right}}{character string}

\item{\code{inRearrangement.right}}{a \code{LogicalList}}
}}
\examples{
ExonSubset()
}
\keyword{-internal}

