% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges-utils.R
\name{expandGRanges}
\alias{expandGRanges}
\title{Expand a genomic interval}
\usage{
expandGRanges(granges, size)
}
\arguments{
\item{granges}{a \code{GRanges} object}

\item{size}{a length-one numeric vector specifying the number of
basepairs to subtract from the start and add to the ends of the
intervals in object \code{granges}}
}
\description{
Expand a genomic interval by a user-specified number of basepairs.
}
\examples{
library(GenomeInfoDb)
library(BSgenome.Hsapiens.UCSC.hg19)
library(GenomicRanges)
si <- keepSeqlevels(seqinfo(BSgenome.Hsapiens.UCSC.hg19), "chr1")
gr <- GRanges("chr1", IRanges(15e3, 20e3), seqinfo=si)
gr2 <- expandGRanges(gr, 5e3)
## will not go beyond the seqlengths
gr3 <- GRanges("chr1", IRanges(seqlengths(gr)-5000, seqlengths(gr)-2000),
               seqinfo=si)
gr4 <- expandGRanges(gr3, 10e3)
end(gr4) == seqlengths(gr) 


}
