% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-utils.R
\name{writeUnmappedToFasta}
\alias{writeUnmappedToFasta}
\alias{writeToFasta}
\title{Helper function to write tags to FASTA files}
\usage{
writeUnmappedToFasta(tags, fasta.file)

writeToFasta(tags, fasta.file)
}
\arguments{
\item{tags}{a data.frame of reads or tags to re-align by BLAT}

\item{fasta.file}{character string providing complete path to a FASTA file}
}
\description{
Helper function to write tags to FASTA files
}
