% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-utils.R
\name{fasta_unmapped}
\alias{fasta_unmapped}
\title{Write the read sequences of unmapped-mapped read pairs to disk in fasta format}
\usage{
fasta_unmapped(dp, unmapped_read, id)
}
\arguments{
\item{dp}{a \code{DataPaths} object}

\item{unmapped_read}{a \code{GRanges} object}

\item{id}{a length-one character vector for the sample id}
}
\value{
nothing is returned
}
\description{
Write the read sequences of unmapped-mapped read pairs to disk in fasta format
}
\seealso{
See \code{\link{unmapped_read}} for extracting a
  \code{GRanges} representation of unmapped-mapped read pairs where
  the mapped read overlap with a set of query genomic intervals,
  such as candiate rearrangements.
}
