% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\docType{methods}
\name{filterBy}
\alias{filterBy}
\alias{filterBy,GRanges,GRanges-method}
\title{Remove the genomic intervals in query that overlap with intervals
in subject}
\usage{
filterBy(query, subject, type = "any", ...)

\S4method{filterBy}{GRanges,GRanges}(query, subject, type = "any", ...)
}
\arguments{
\item{query}{a \code{\linkS4class{BamViews}} or \code{\linkS4class{GRanges}} instance}

\item{subject}{a \code{\linkS4class{GRanges}} object}

\item{type}{see \code{\link{findOverlaps}}}

\item{...}{Additional arguments passed to \code{findOverlaps}}
}
\value{
Returns the \code{query} with intervals that overlap \code{subject} removed
}
\description{
Remove the genomic intervals in query that overlap with intervals
in subject
}
\seealso{
\code{\link[GenomicRanges]{findOverlaps}}
}
