% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{filterGermlineRear}
\alias{filterGermlineRear}
\title{Applies germline CNV and sequence-based filters for non-redundant
identification of germline rearrangements}
\usage{
filterGermlineRear(object, params, germline_filters,
  return_stats = FALSE)
}
\arguments{
\item{object}{A \code{RearrangementList} object}

\item{params}{A \code{RearrangementParams} object}

\item{germline_filters}{A \code{GRanges} object of germline CNVs and
sequence-based filters}

\item{return_stats}{length-one logical vector. For internal use only}
}
\description{
Applies germline CNV and sequence-based filters for non-redundant
identification of germline rearrangements
}
\seealso{
See \code{\link{filterRearExperiment}} for a wrapper that
  saves filtered results. See \code{\link{filterRearrangementList}} for
  filtering somatic rearrangements.
}
