% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{filterGermlineRearList}
\alias{filterGermlineRearList}
\title{Exclude linked tag cluster that overlap with germline CNVs or other
sequence artifacts}
\usage{
filterGermlineRearList(x, dirs, rp = RearrangementParams(), germline_filters)
}
\arguments{
\item{x}{a \code{RearrangementList}}

\item{dirs}{a \code{DataPaths} object}

\item{rp}{a \code{RearrangementParams} object}

\item{germline_filters}{a \code{GRanges} object of germline CNV and sequence filters}
}
\description{
This function is intended for internal use, primarily for
 delineating rearrangements that may occur in the germline.
}
\keyword{internal}
