% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{filterPairedReads}
\alias{filterPairedReads}
\title{Remove duplicates paired r}
\usage{
filterPairedReads(gpairs, bins, params)
}
\arguments{
\item{gpairs}{TODO}

\item{bins}{TODO}

\item{params}{TODO}
}
\description{
Remove duplicates paired r
}
