% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{filterRear}
\alias{filterRear}
\title{Filter candidate rearrangements for germline}
\usage{
filterRear(rdata, params = RearrangementParams())
}
\arguments{
\item{rdata}{a list of rearrangements and filters}

\item{params}{a \code{RearrangementParams} object}
}
\description{
Filter candidate rearrangements for germline
}
