% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{filterRearExperiment}
\alias{filterRearExperiment}
\title{Exclude linked tag cluster that overlap with germline CNVs or other
sequence artifacts}
\usage{
filterRearExperiment(
  x,
  dirs,
  cnvs,
  germline_filters,
  rear_filter,
  rp = RearrangementParams()
)
}
\arguments{
\item{x}{a \code{RearrangementList}}

\item{dirs}{a \code{DataPaths} object}

\item{cnvs}{a list of deletions and amplicons}

\item{germline_filters}{a \code{GRanges} object of germline filters}

\item{rear_filter}{a \code{GRanges} object of rearrangements found in germline}

\item{rp}{a \code{RearrangementParams} object}
}
\value{
a \code{RearrangementList}
}
\description{
This is a wrapper for \code{filterRearrangementList} (somatic) and
\code{filterGermlineRear} (germline).  The filtered rearrangements
are saved to an intermediate file.  If the file already exists, the
saved result is read from disk.  To rerun the filtering step, the
intermediate files must be removed.  See examples.
}
\details{
REFACTORING: Too many arguments. Unclear distinction between
germline_filters and rear_filter.  Should have a separate function for
germline, or have an S4 germline class.



The amplicons (amps) and deletions (del.gr) will be subset by the
  mcols variable 'id'.
}
\seealso{
See \code{filterRearrangementList} for filtering somatic
  rearrangements and \code{filterGermlineRear} for filtering
  germline rearrangements.  See \code{\link{listCNVs}} for listing
  the deletions and amplicons for a given sample.
}
