% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{filterRearrangementList}
\alias{filterRearrangementList}
\title{Applies several filters for somatic rearrangements}
\usage{
filterRearrangementList(object, params, amps, del.gr, rear_filter,
  germline_filters = GRanges(), return_stats = FALSE)
}
\arguments{
\item{object}{A \code{RearrangementList}}

\item{params}{A \code{RearrangementParams} object}

\item{amps}{\code{GRanges} of amplicons}

\item{del.gr}{\code{GRanges} of deletions}

\item{rear_filter}{\code{GRanges} object of germline rearrangements}

\item{germline_filters}{\code{GRanges} of germline CNVs and sequence filters}

\item{return_stats}{a length-one logical vector.  For internal use.}
}
\description{
Removes rearrangement intervals that overlap with previously
identified somatic amplicons and deletions, as well as germline
CNVs and rearrangements.
}
