% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{fiveTo3List}
\alias{fiveTo3List}
\title{Reorders each rearrangement as two 5-prime to 3-prime orientations}
\usage{
fiveTo3List(rlist, build)
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{build}{character string providing genome build (only hg19 or hg18 are supported)}
}
\value{
a \code{RearrangementList}
}
\description{
Reorders each rearrangement as two 5-prime to 3-prime orientations
}
\examples{
 library(trellis)
 extdata <- system.file("extdata", package="svbams")
 rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
 rlist <- readRDS(rfile)
 near.coding <- seqJunctionNearTx(rlist, "hg19")
 \dontrun{
   rlist2 <- fiveTo3List(rlist[near.coding], build="hg19")
   rlist2
}
}
