% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{fiveTo3Prime}
\alias{fiveTo3Prime}
\title{Put the linked genomic intervals in 5-prime to 3-prime order with respect to the rearranged genome}
\usage{
fiveTo3Prime(r, build, maxgap = 5000)
}
\arguments{
\item{r}{a \code{Rearrangement} object}

\item{build}{string providing build of reference genome ('hg18', 'hg19', or 'hg38')}

\item{maxgap}{maximum distance between read and transcript for transcript to be considered overlapping}
}
\description{
Put the linked genomic intervals in 5-prime to 3-prime order with respect to the rearranged genome
}
\examples{
  extdata <- system.file("extdata", package="svbams")
  rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
  rlist <- readRDS(rfile)
  r <- rlist[[1]]
  r2 <- fiveTo3Prime(r, "hg19")
  df <- rearDataFrame(r2[[1]], "hg19")
  ggRearrange(df)
}
\seealso{
\code{\link{rearDataFrame}} \code{\link{ggRearrange}}
}
