% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{focalAmpliconDupRanges}
\alias{focalAmpliconDupRanges}
\title{Identify lower copy focal amplicons (possibly duplicatons) not already
annotated as amplicons and not contained within the germline filters}
\usage{
focalAmpliconDupRanges(object, params)
}
\arguments{
\item{object}{A \code{AmpliconGraph}}

\item{params}{a list of parameters such as that generated by \code{ampliconParams}.}
}
\value{
a reduced \code{GRanges} of lower-copy amplicons
}
\description{
Select all duplicated segments (see \code{isDuplication}) less than `maxgap`
kb from the seed amplicons of the graph that are not within a germline CNV or
outlier. Germline CNVs and outliers can be conveniently extracted from the
graph object as given in the examples.
}
\examples{
  ## load a previously saved AmpliconGraph from a regression unit test
  \dontrun{
    path <- system.file("testthat", package="trellis")
    ag <- readRDS(file.path(path, "addFocalDups.ffab104.rds"))
    params <- ampliconParams()
    focalAmpliconDupRanges(ag, params)
}
}
\seealso{
\code{\link{ampliconParams}}
}
