% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrangement-class.R, R/RearrangementList-class.R
\docType{methods}
\name{fractionLinkingTags}
\alias{fractionLinkingTags}
\alias{fractionLinkingTags,Rearrangement,ANY-method}
\alias{fractionLinkingTags,Rearrangement-method}
\alias{fractionLinkingTags,RearrangementList-method}
\alias{fractionLinkingTags<-}
\alias{fractionLinkingTags<-,Rearrangement-method}
\title{The fraction of improper read pairs that link two clusters of improper reads}
\usage{
fractionLinkingTags(object)

fractionLinkingTags(object) <- value

\S4method{fractionLinkingTags}{Rearrangement}(object)

\S4method{fractionLinkingTags}{Rearrangement}(object) <- value

\S4method{fractionLinkingTags}{RearrangementList}(object)
}
\arguments{
\item{object}{a \code{Rearrangement} object}

\item{value}{length-one numeric vector}
}
\value{
numeric
}
\description{
For bona fide deletions, inversions, and translocations, we would
expect that the fraction of improper read pairs that link two
clusters of improper reads to be high.  For amplicons, the fraction
of reads linking two clusters may be small.
}

