% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{fullTranscripts}
\alias{fullTranscripts}
\title{List full transcripts for genes spanning a rearrangement}
\usage{
fullTranscripts(object)
}
\arguments{
\item{object}{a \code{Transcripts} object}
}
\value{
a \code{GRangesList}
}
\description{
List full transcripts for genes spanning a rearrangement
}
\examples{
data(rear_cds, package="trellis")
fullTranscripts(rear_cds)
}
\seealso{
\code{\link{inFrameFusions}}
}
