% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fuse}
\alias{fuse}
\title{Fuse transcripts}
\usage{
fuse(object, nms)
}
\arguments{
\item{object}{a \code{Transcripts}-derived class}

\item{nms}{name of transcript (character-string)}
}
\description{
Fuse transcripts
}
