% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{fuseCDS_Rlist}
\alias{fuseCDS_Rlist}
\title{Extract the CDS involved in each rearrangement of a RearrangementList object}
\usage{
fuseCDS_Rlist(rlist, jxns)
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{jxns}{a \code{GRanges} specifying the 5-prime and 3-prime genomic regions that form a new sequence junction}
}
\value{
a \code{List} of the CDS
}
\description{
Extract for each rearrangement the full CDS of the fused sequence in the somatic genome (fusions), the partial CDS of the 5-prime (tum.5p) and 3-prime (tum.3p) transcripts that are involved in the fusion, and the full CDS of the 5-prime (ref.5p) and 3-prime transcripts in the reference genome.
}
