% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{fusionData}
\alias{fusionData}
\title{Collect transcript and protein-level data supporting a fusion for into a list object for subsequent plotting}
\usage{
fusionData(data.list, build = c("hg19", "hg18"))
}
\arguments{
\item{data.list}{a list object created by \code{listFusionData}}

\item{build}{length-one character vector providing genome build}
}
\value{
a named list
}
\description{
Collect transcript and protein-level data supporting a fusion for into a list object for subsequent plotting
}
