% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{fusionExperiment}
\alias{fusionExperiment}
\title{Evaluates rearrangements for in-frame fusions}
\usage{
fusionExperiment(file, id, rlist, dirs, txdb, orgdb, genome)
}
\arguments{
\item{file}{file path}

\item{id}{sample id (character string)}

\item{rlist}{a \code{RearrangementList}}

\item{dirs}{a \code{DataPaths} object}

\item{txdb}{a \code{TxDb} object}

\item{orgdb}{an \code{OrgDb} object}

\item{genome}{aa \code{BSgenome} object}
}
\value{
a list of \code{tbl_df} objects
}
\description{
Saves the results of fusionTable to disk.  If file already exists,
only read in the file and return a list of \code{tbl_df} objects.
}
\seealso{
\code{fusionTable}
}
