% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{fusionList}
\alias{fusionList}
\title{A title}
\usage{
fusionList(rlist, id = "", build = c("hg19", "hg18"))
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{id}{character string providing identifier for RearrangementList}

\item{build}{genome build ('hg18', 'hg19', or 'hg38')}
}
\description{
A title
}
