% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{ga2gr}
\alias{ga2gr}
\title{Convert GAlignmentPairs to GRanges while maintaining read pair information}
\usage{
ga2gr(ga, is.improper = FALSE)
}
\arguments{
\item{ga}{a \code{GAlignments} object}

\item{is.improper}{a length-one logical vector.  \code{TRUE} if the reads are improperly paired, \code{FALSE} otherwise.}
}
\value{
A \code{GRanges} object with metadata columns containing read pair information.
}
\description{
Melts \code{GAlignmentPairs} objects to \code{GRanges} objects.
}
\examples{
library(svbams)
path <- system.file("extdata", package="svbams")
bam.file <- file.path(path, "cgov10t.bam")
irp <- getImproperAlignmentPairs(bam.file)
igr <- ga2gr(irp, is.improper=TRUE)
prp <- getProperAlignmentPairs(bam.file)
pgr <- ga2gr(prp, is.improper=FALSE)
 
}
