% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAlignmentPairs-utils.R
\name{.GAlignmentPairs}
\alias{.GAlignmentPairs}
\title{Constructor for an empty GAlignmentPairs object}
\usage{
.GAlignmentPairs()
}
\value{
an empty \code{GAlignmentPairs} object
}
\description{
Constructor for an empty GAlignmentPairs object
}
\keyword{internal}

