% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{germlineFilters}
\alias{germlineFilters}
\title{Identify focal, somatic hemizygous deletions and somatic homozygous
deletions}
\usage{
germlineFilters(preprocess, germline_filters, param = DeletionParam())
}
\arguments{
\item{preprocess}{a list of preprocessed data. See \code{preprocessData}}

\item{germline_filters}{A \code{GRanges} object (e.g., germline
CNVs and regions of low sequence quality)}

\item{param}{A \code{DeletionParam} object}
}
\value{
a named \code{GRanges} object.  The names are given by
  \code{paste0("sv", seq_along(g))} where g is the \code{GRanges}
  object.
}
\description{
Segmentation of bin-level normalized log2 ratios yields a set of genomic
intervals each having an associated mean. We define candidate somatic
deletions as those having a segment mean less than a specified cutoff (e.g.,
the theoretical segment mean of a hemizygous deletion on log2 scale). For
each candidate CNV, we assess whether the variant can be explained by events
identified in germline samples processed in the same batch, or by sequence
artifacts (low mappability and/or GC). To help ensure that the deletion is
focal (not a small deletion embedded in a larger deletion), we (i) verify
that the candidate CNV (\code{cnv}) has a mean log ratio less than the mean
log ratio of the neighboring segments by an amount of approx. -0.5
[log2(0.7)] and (ii) assess whether the interval could plausibly be a large
hemizygous deletion. Candidate somatic deletions identified by this function
have the following properties: (i) not germline, (ii) unlikely to be large
hemizygous deletions, (iii) the difference in mean of the segment and
neighboring segments is less than -0.5, and (iv) have a width of at least
2kb.
}
\seealso{
\code{\link{preprocessData}}
}
