% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{get_improper_readpairs}
\alias{get_improper_readpairs}
\title{Extract all improperly paired reads from an object with queryRanges defined}
\usage{
get_improper_readpairs(object, bam.file)
}
\arguments{
\item{object}{An object for which \code{queryRanges} method is defined}

\item{bam.file}{The complete file path to a bam file.}
}
\description{
This function is a wrapper for readGAlignments followed by
makeGAlignmentPairs2.  Only read pairs in which both the first read
and the last read in a pair overlap with the queryRanges are
returned.  Note, a given read pair need not overlap the same
queryRange.  To allow fuzzy matching of alignments to the
queryRanges, the queryRanges are expanded by 2kb in each direction.
}
