% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{get_readpairs}
\alias{get_readpairs}
\title{Parse BAM file for improper read pairs near a set of GRanges}
\usage{
get_readpairs(object, bam.file, flags = scanBamFlag())
}
\arguments{
\item{object}{Typically an \code{AmpliconGraph}, though the only
requirement is that the method \code{queryRanges} is defined}

\item{bam.file}{character-vector providing valid complete path to a
bam file}

\item{flags}{length-two integer vector as given by \code{scanBamFlags}}
}
\value{
A \code{GAlignmentPairs} object
}
\description{
All reads aligned to the intervals given by
\code{queryRanges(object)} are identified by the low-level function
\code{.scan_all_readpairs}.  This function reads alignments by
\code{readGAlignments} and then makes pairs of the alignments by
\code{makeGAlignmentPairs2}.  The latter function is an adaption of
the function \code{makeGAlignmentPairs} implemented in the
\code{GenomeAlignments} package but allows for the read pairs to be
improper.
}
