% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{get_readpairs2}
\alias{get_readpairs2}
\title{Extract reads from a bam file}
\usage{
get_readpairs2(g, bam.file, flags = scanBamFlag())
}
\arguments{
\item{g}{A \code{GRanges} object}

\item{bam.file}{A character vector of the path to a bam file}

\item{flags}{A length-two integer vector as given by \code{scanBamFlags}}
}
\value{
A \code{GAlignmentPairs} object
}
\description{
Parses a BAM file for read pairs near intervals specified by a \code{GRanges} object and 
returns the read pairs in a \code{GAlignmentPairs} object.
}
\seealso{
get_readpairs
}
