% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{getCDS}
\alias{getCDS}
\title{Finds all genes overlapping a rearrangement (ignoring strand) and
returns the full CDS of each}
\usage{
getCDS(rear, tx, cds)
}
\arguments{
\item{rear}{A \code{Rearrangement} object}

\item{tx}{A \code{GRanges} object of transcripts from the \code{TxDb} object}

\item{cds}{A \code{GRangesList} object containing the CDS for each transcript}
}
\value{
a \code{Transcripts} object
}
\description{
Finds all genes overlapping the \code{linkedBins} of a
\code{Rearrangement} object (ignoring strand) and extracts the full
CDS.  The rearrangement type inferred from the position and strand
orientation of read pairs (see \code{RearrangementType}) determines
the orientation of the genes in the rearranged genome.  See detail
for the generic symbols used to denote a fused transcript.
}
\details{
#' \preformatted{
Reference   (-- chr1, == chr2,  .... not in rearranged transcript)
          A                                    C
5' + ------------|.........   ...............|==============
3' - ------------|.........   ...............|==============
          B                                    D


Tumor

      chr1  A        C
5' + -------------|==========
3' - -------------|==========
            B        D

}

\strong{Deletions:} 
For a deletion, we have chr1 = chr2, R1+ < R2-, and R1- > R2+.  The
  possible fusions are AC and DB.

\strong{Translocations:} A read pair in a translocation supports
  one of \code{trans1}, \code{trans2}, \code{trans3}, and
  \code{trans4}.  For a balanced translocation, all four types
  could be observed.  We analyze the data as a balanced
  translocation and evaluate all possible fusions even if not all
  types are observed.  Again, the possible fusions are AC and DB.

Amplicons and translocations in which the sequences is inverted in
  the rearranged genome are analyzed as inversions.

\strong{Inversions:} There are 8 possible fusions indicated by an
  inverted read pair denoted as \code{invX}, where X = 1-8.
  Possible fused gene products are AD, CB, BC, and DA.
}
\examples{
  options(warn=-1)
  library("TxDb.Hsapiens.UCSC.hg19.refGene")
  txdb <- TxDb.Hsapiens.UCSC.hg19.refGene
  tx <- transcripts(txdb)
  cds.all <- cdsBy(txdb, "tx", use.names=TRUE)
  data(rear_list, package="trellis")
  r <- rear_list[["18557-18736"]]
  ##data(rear_cds, package="trellis")
  ##trace(getCDS, browser)
  rear_cds <- getCDS(r, tx, cds.all)
  fusions(rear_cds)
  ## There are for transcripts for gene 'B'.  The accessor txB
  ## returns the CDSs for each of the transcripts as a
  ## \\code{GRangesList}
  txB(rear_cds)
  ## Similarly, there are accessors \\code{txA}, \\code{txC}, and
  ## \\code{txD}
}
