% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{getProperAlignmentPairs}
\alias{getProperAlignmentPairs}
\title{Extract properly paired reads from a bam file}
\usage{
getProperAlignmentPairs(bam.file, param = properAlignmentParams(mapqFilter =
  0))
}
\arguments{
\item{bam.file}{a \code{BamViews} object}

\item{param}{a \code{ScanBamParam} object.}
}
\value{
a \code{GAlignmentPairs} object
}
\description{
Function used by rearrangement analysis to extract the sequence of
properly paired reads from a bam file.
}
\examples{
library(svbams)
path <- system.file("extdata", package="svbams")
bam.file <- file.path(path, "cgov10t.bam")
irp <- getProperAlignmentPairs(bam.file)
  
}
\seealso{
See \code{\link{properAlignmentParams}} for creating a
  \code{ScanBamParam} object with the appropriate flags for extracting
  proper read pairs.
}
\keyword{internal}
