% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{getSequenceOfReads_bedops}
\alias{getSequenceOfReads_bedops}
\title{getSequenceOfReads for CRAM}
\usage{
getSequenceOfReads_bedops(rlist.bedops, MAX = 25, sample_name = sample)
}
\arguments{
\item{rlist.bedops}{}

\item{MAX}{}

\item{sample}{}
}
\value{
Tagged sequences from BLAT
}
\description{
getSequenceOfReads for CRAM
}
