% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{ggRearrange}
\alias{ggRearrange}
\title{ggplot wrapper for plotting reads supporting a rearrangement}
\usage{
ggRearrange(df, ylab = "Read pair index", basepairs = 400, num.ticks = 5)
}
\arguments{
\item{df}{a \code{data.frame} as created by \code{rearDataFrame}}

\item{ylab}{y-axis label}

\item{basepairs}{integer specifying size of window in basepairs}

\item{num.ticks}{integer specifying number of x-axis ticks}
}
\description{
ggplot wrapper for plotting reads supporting a rearrangement
}
\examples{
  extdata <- system.file("extdata", package="trellis")
  rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
  rlist <- readRDS(rfile)
  r <- rlist[[1]]
  df <- rearDataFrame(r, "hg19")
  ggRearrange(df)
}
\seealso{
\code{\link{rearDataFrame}}
}
