% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess-class.R
\name{granges_copynumber}
\alias{granges_copynumber}
\title{Compute the mean normalized read-depth for a set of intervals}
\usage{
granges_copynumber(gr, pviews)
}
\arguments{
\item{gr}{a \code{GRanges} object}

\item{pviews}{a \code{PreprocessViews2} object}
}
\description{
Calculates the mean normalized read-depth for a set of genomic
intervals in a \code{GRanges} object.
}
\keyword{internal}
