% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{groupAmplicons}
\alias{groupAmplicons}
\title{Assign a group id for nodes that are linked}
\usage{
groupAmplicons(gN)
}
\arguments{
\item{gN}{a \code{AmpliconGraph} object}
}
\description{
Takes a list of amplicons (GRanges object) and an undirected graph
representation of the amplicons
}
