% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R
\docType{methods}
\name{groupedVariant}
\alias{groupedVariant}
\alias{groupedVariant<-}
\alias{groupedVariant,StructuralVariant-method}
\alias{groupedVariant<-,StructuralVariant-method}
\alias{groupedVariant,StructuralVariant,ANY-method}
\title{Accessor for variable indicating a grouping for deletions}
\usage{
groupedVariant(object)

groupedVariant(object) <- value

\S4method{groupedVariant}{StructuralVariant}(object)

\S4method{groupedVariant}{StructuralVariant}(object) <- value
}
\arguments{
\item{object}{a \code{StructuralVariant}}

\item{value}{a factor indicating a grouping for deletions}
}
\value{
factor indicating a grouping for deletions
}
\description{
The grouping can indicate, for example, overlapping hemizygous
deletions.
}
\keyword{internal}
