% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improper-methods.R
\docType{methods}
\name{improper}
\alias{improper}
\alias{improper<-}
\alias{improper,Rearrangement-method}
\alias{improper<-,StructuralVariant,GAlignmentPairs-method}
\alias{improper,StructuralVariant,GAlignmentPairs-method}
\alias{improper<-,Rearrangement,GAlignmentPairs-method}
\alias{improper,Rearrangement,GAlignmentPairs-method}
\title{Accessor for improper read pairs}
\usage{
improper(object)

improper(object) <- value

\S4method{improper}{Rearrangement}(object)

\S4method{improper}{StructuralVariant,GAlignmentPairs}(object) <- value

\S4method{improper}{Rearrangement,GAlignmentPairs}(object) <- value
}
\arguments{
\item{object}{a \code{Rearrangement} or \code{StructuralVariant} object}

\item{value}{a \code{GAlignmentPairs} object}

\item{value}{a \code{GAlignmentPairs} object}
}
\description{
Extract the improperly paired reads.  Reads are flagged as improper
by the alignment algorithm.  Typically, this indicates that the
separation between a read and its mate is larger than expected, the
orientation of the read and its mate is different from that
anticipated in the reference genome, or the reads align to
different chromosomes.
}
\seealso{
\code{\linkS4class{Rearrangement}}
}
\keyword{internal}
