% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Deletion-class.R
\docType{methods}
\name{indexProper}
\alias{indexProper}
\alias{indexProper<-}
\alias{indexImproper<-}
\alias{indexImproper}
\alias{indexProper<-,StructuralVariant,list-method}
\alias{indexProper,StructuralVariant,list-method}
\alias{indexImproper<-,StructuralVariant,list-method}
\alias{indexImproper,StructuralVariant,list-method}
\alias{indexProper,StructuralVariant-method}
\alias{indexProper<-,StructuralVariant,list-method}
\alias{indexProper,StructuralVariant,list-method}
\alias{indexImproper,StructuralVariant-method}
\alias{indexImproper<-,StructuralVariant,list-method}
\alias{indexImproper,StructuralVariant,list-method}
\title{Accessor of an index for the proper read pairs}
\usage{
indexProper(object)

indexProper(object) <- value

indexImproper(object) <- value

indexImproper(object)

\S4method{indexProper}{StructuralVariant,list}(object) <- value

\S4method{indexImproper}{StructuralVariant,list}(object) <- value

\S4method{indexProper}{StructuralVariant}(object)

\S4method{indexProper}{StructuralVariant,list}(object) <- value

\S4method{indexImproper}{StructuralVariant}(object)

\S4method{indexImproper}{StructuralVariant,list}(object) <- value
}
\arguments{
\item{object}{a \code{StructuralVariant}}

\item{value}{a list}
}
\value{
a numeric vector
}
\description{
Accessor of an index for the proper read pairs
}
\keyword{internal}
