% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Preprocess-class.R
\docType{methods}
\name{indexRanges}
\alias{indexRanges}
\alias{indexRanges<-}
\alias{indexRanges,PreprocessViews2-method}
\alias{indexRanges<-,PreprocessViews2-method}
\title{Indexes the GRanges of a BamViews-derived class}
\usage{
indexRanges(object)

indexRanges(object) <- value

\S4method{indexRanges}{PreprocessViews2}(object)

\S4method{indexRanges}{PreprocessViews2}(object) <- value
}
\arguments{
\item{object}{a \code{BamViews}-derived object}

\item{value}{an integer-vector}
}
\description{
Indexes the GRanges of a BamViews-derived class
}
\keyword{internal}
