% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{inFrameFusions}
\alias{inFrameFusions}
\title{Evaluates whether a rearranged DNA sequence is in-frame}
\usage{
inFrameFusions(fused.proteins, ref.protein, fused.txlist)
}
\arguments{
\item{fused.proteins}{the rearranged protein sequence}

\item{ref.protein}{the unrearranged (reference) protein sequence}

\item{fused.txlist}{a \code{GRangesList} of the fused transcripts}
}
\description{
A rearranged sequence is considered in-frame if the protein sequence derived
from the DNA sequence 3-prime of the new sequence junction is a subset of the
reference protein sequence.
}
\examples{
  library(BSgenome.Hsapiens.UCSC.hg19)
  genome <- BSgenome.Hsapiens.UCSC.hg19
  data(rear_cds, package="trellis")
  unrear_cds <- fullTranscripts(rear_cds)
  ##
  ## referenceProtein
  ##
  ref_protein <- referenceProtein(genome, unrear_cds, names(unrear_cds))
  fused_tx <- fuse(clip(rear_cds))
  fused_protein <- tumorProtein(genome, fused_tx)
  inFrameFusions(fused_protein, ref_protein, fused_tx)
}
\seealso{
See \code{\link{referenceProtein}} and \code{\link{tumorProtein}}
  for extracting the unrearranged and rearranged protein sequences,
  respectively. See \code{\link{clip}} and
  \code{\link{fuse}} for extracting the fused, clipped transcript
  sequence as a \code{GRangesList} object. See \code{\link{fullTranscripts}}
  for extracting the full transcripts as a \code{GRangesList}.
}
