% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{inFrameList}
\alias{inFrameList}
\title{Evaluate whether fusion amino acid sequence is a subsequence of the reference}
\usage{
inFrameList(fusion.frames, ref.frames)
}
\arguments{
\item{fusion.frames}{a \code{List} created by \code{partitionAASequence}}

\item{ref.frames}{a \code{List} created by \code{organizeReferenceByFrame}}
}
\description{
For each fusion and for each possible read frame, we check whether the amino acid sequence derived from the 5- and 3-prime transcripts of the fusion are subsequences of the full amino acid sequence of the reference genome translated in the same frame
}
\seealso{
\code{\link{partitionAASequence}} \code{\link{organizeReferenceByFrame}}
}
