% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{intersectionAligned}
\alias{intersectionAligned}
\title{Compute the intersection of the split read alignment}
\usage{
intersectionAligned(g)
}
\description{
Example
}
\details{
split 1     :  -----------
split 2     :           -----------
intersection:  2
}
