% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges-utils.R
\name{intOverWidth}
\alias{intOverWidth}
\title{Find the fraction of genomic interval overlapping with another
genomic interval.}
\usage{
intOverWidth(query, subject)
}
\arguments{
\item{query}{a \code{GRanges} object}

\item{subject}{a \code{GRanges} object}
}
\value{
A numeric vector of the same length as the query
  \code{GRanges}
}
\description{
Compute the fraction of a query genomic interval that overlaps with another
set of genomic intervals. We divide the total width of the intersection with
the width of the query.
}
\examples{
 library(GenomicRanges)
 autosomes <- function() paste0("chr", 1:22)
 del <- GRanges(c("chr1", "chr2"), IRanges(c(1, 1), c(10, 10)))
 filters <- GRanges(c("chr1", "chr2"), IRanges(c(5, 8), c(20, 10000)))
 test <- intOverWidth(del, filters)
 identical(test, c(0.6, 0.3))

}

