% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{is_valid_splits}
\alias{is_valid_splits}
\title{Check that the split reads available for each rearrangement span both sides of the sequence junction}
\usage{
is_valid_splits(rlist, maxgap = 100)
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{maxgap}{length-one numeric vector passed to \code{subsetByOverlaps}}
}
\value{
logical vector of the same length as \code{rlist}
}
\description{
Check that the split reads available for each rearrangement span both sides of the sequence junction
}
\examples{

}
