% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{isComplex}
\alias{isComplex}
\title{Ad-hoc assessment of the complexity of a rearrangement using the strand of the supporting read pairs}
\usage{
isComplex(x)
}
\arguments{
\item{x}{a \code{Rearrangement} or \code{RearrangementList}}
}
\description{
For most rearrangements, the strand of the 5-prime and 3-prime reads belonging to a rearranged read pair will be consistent. For example, all rearranged read pairs are +/-, indicating a fusion on the positive strand (R1+, R2-) or negative strand (R1-, R2+).
}
\examples{
  extdata <- system.file("extdata", package="svrearrange")
  r <- readRDS(file.path(extdata, "cgov1t_complex_rearrangement.rds"))
  s <- type(r)
  print(s)
  isComplex(r)
}
