% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{isDuplicate}
\alias{isDuplicate}
\title{Assesses whether any read pairs are duplicates in a GAlignmentPairs object}
\usage{
isDuplicate(aln.pairs)
}
\arguments{
\item{aln.pairs}{a \code{GAlignmentPairs} object}
}
\value{
logical vector of same length as the \code{aln.pairs}
}
\description{
Assesses whether any read pairs are duplicates in a GAlignmentPairs object
}
