% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{isNotGermline}
\alias{isNotGermline}
\title{Assess whether a region can be attributable to a germline CNV or
artifact seen in germline samples}
\usage{
isNotGermline(g, all_filters, param = DeletionParam())
}
\arguments{
\item{g}{a \code{GRanges} object (e.g., candidate deletions)}

\item{all_filters}{a \code{GRanges} object (e.g., deletions,
amplifications, and outliers identified in the germline)}

\item{param}{an instance of \code{DeletionParam}}
}
\value{
a locical vector having the same length as \code{g}
}
\description{
In our application, \code{all_filters} is a \code{GRanges} object
comprised of germline CNVs and sequence filters.  The latter is a
collection of 1kb bins (reduced) that have low mappability (< 0.75)
and/or low GC content (< 0.1).  For each interval in the candidate
somatic \code{GRanges} object (\code{g}), we compute the fraction
of the interval spanned by the germline filters.  If the fraction
of the candidate somatic variant spanned by the germline filter is
less than \code{max_proportion_in_filter} AND the width not spanned
by the germline filters is less than \code{min_width}, this
function evaluates to \code{TRUE}.
}
\seealso{
See \code{DeletionParam} for default values of
  \code{max_proportion_in_filter} and \code{min_width}.
}
