% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{joinNearGRanges}
\alias{joinNearGRanges}
\title{Merge adjacent amplicons that have a similar segment mean}
\usage{
joinNearGRanges(object, params)
}
\arguments{
\item{object}{a \code{GRanges} object}

\item{thr}{a length-one numeric vector. If the difference of the average log
ratio for two segments is less than this number, combine the segments. The
segment mean of the merged GRanges will be the weighted average, where the
weights are the segment widths.}

\item{MIN_WIDTH}{minimum size of amplicon (default 2000)}
}
\value{
a \code{GRanges} object of the merged segments
}
\description{
Merges adjacent amplicons (<= 3kb separation) that have similar
segment means.
}
\details{
Merging is performed by the \code{reduce} operation.
}
\keyword{internal}
