% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchTrellis.R
\name{launchTrellis}
\alias{launchTrellis}
\title{Launch the Trellis shiny app}
\usage{
launchTrellis()
}
\arguments{
\item{corpus.path}{Path to directory containing separate text files.}

\item{k}{Number of topics to train}

\item{out.path}{Path to output location for the Trellis model. If NULL, ignored.}

\item{ldavis.data.path}{Path to output location for term.frequency and doc.length. If NULL, ignored. See \code{trellis::toLDAvis}.}
}
\description{
TODO(tfs): documentation in R doc. DOCUMENT what the package is meant to do
           Point to instructional materials for curation of text corpus. See ajbc slides
			  BUild function to separate out file contents into individual text files.
           Add little badges to github
}
\examples{
\dontrun{
    TODO
}
}
\seealso{
\pkg{\link{stm}}

\link[dest=https://ldavis.cpsievert.me/reviews/reviews.html]{ldavis}
}
