% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrangement-class.R,
%   R/RearrangementList-class.R
\docType{methods}
\name{linkedBins}
\alias{linkedBins}
\alias{linkedBins<-}
\alias{linkedBins,Rearrangement-method}
\alias{linkedBins<-,Rearrangement-method}
\alias{,}
\alias{linkedBins,Rearrangement,ANY-method}
\alias{linkedBins,RearrangementList-method}
\alias{linkedBins<-,RearrangementList-method}
\alias{linkedBins,RearrangementList,ANY-method}
\title{Accessor for clusters of reads belonging to improper read pairs that are linked.}
\usage{
linkedBins(object)

linkedBins(object) <- value

\S4method{linkedBins}{Rearrangement}(object)

\S4method{linkedBins}{Rearrangement}(object) <- value

\S4method{linkedBins}{RearrangementList}(object)

\S4method{linkedBins}{RearrangementList}(object) <- value
}
\arguments{
\item{object}{a \code{Rearrangement} object}

\item{value}{a \code{GRanges} object with a column \code{linked.to}
in the element metadata.}
}
\description{
Clusters of improper reads that are linked by the read pairs are
suggestive of a new (somatic) sequence junction not present in the
reference genome.  The linking is represented by a \code{GRanges}
object with an interval demarcating the boundaries of a cluster of
improper tags.  The tag cluster that is linked by the pairing
information is put in the element metadata (\code{mcols}) of the
\code{GRanges} object.
}
