% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{linkedTo}
\alias{linkedTo}
\title{A region linked by improperly paired reads}
\usage{
linkedTo(x)
}
\arguments{
\item{g}{a \code{GRanges} object with value `linked.to` in the `mcols`}
}
\description{
A region linked by improperly paired reads
}
