% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCNVs.R
\name{listCNVs}
\alias{listCNVs}
\title{List amplicons and deletions identified for a single sample}
\usage{
listCNVs(amplicon_graph, del_sv)
}
\arguments{
\item{amplicon_graph}{an \code{AmpliconGraph} object}

\item{del_sv}{a \code{StructuralVariant} object from the deletion pipeline}
}
\value{
a list
}
\description{
List amplicons and deletions identified for a single sample
}
