% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{listDeletions}
\alias{listDeletions}
\title{Extract a GRangesList of all identified deletions in a project}
\usage{
listDeletions(path = "data/segment/1deletions", ids)
}
\arguments{
\item{path}{character string providing directory where deletions are saved}

\item{ids}{a character vector of sample identifiers}
}
\value{
a \code{GRangesList} of deletions
}
\description{
Deletions are assumed by be saved in a deletion directory by the name of
the BAM file.
}
