% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPaths-class.R
\name{listDir}
\alias{listDir}
\title{Lists folders for a given topic}
\usage{
listDir(folder, object)
}
\arguments{
\item{folder}{length-one character string}

\item{object}{an object of class \code{DataPaths}}
}
\description{
Lists folders for major topics, together with the path to the major
topic.
}
\details{
TODO: add topics()
}
\examples{
dp <- DataPaths(tempdir(), "Test")
listDir("preprocess", dp)

}
\seealso{
See \code{\linkS4class{DataPaths}}.  See
  \code{\link{topics}} for a listing of the major headings.
}
