% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{listFusionData}
\alias{listFusionData}
\title{Collect fusion-related data into a single list}
\usage{
listFusionData(rlist, fusions)
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{fusions}{an object returned by \code{fusionList}}
}
\value{
a named list
}
\description{
Collect fusion-related data into a single list
}
