% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGenomeFilters.R
\name{listGenomeFilters}
\alias{listGenomeFilters}
\title{Genomic filters for the somatic amplicon analysis}
\usage{
listGenomeFilters(genome = c("hg19", "hg18"))
}
\arguments{
\item{genome}{length-one character vector specifying UCSC genome build}
}
\value{
a named list
}
\description{
When paired normal samples are unavailable, we use germline
lymphoblast cell lines to identify CNVs and outliers that are
common in the germline and unlikely to be somatic.  Filters
relevant for the somatic amplicon analysis include germline CNVs,
centromeric regions, and assembly gaps.  We additionally defined a
'paired bin filter' containing bins that appear to be linked by
improperly spaced reads in the lymphoblast samples.  Aberrantly
spaced linked bins could indicate a new junction in the lymphoblast
cell line or a sequencing artifact.
}
\details{
REFACTORING: Separate filters from the parameter list. Create a single filter
object for both the amplicon and deletion analyses. * See germline_filters.
Do we still need this? *
}
\examples{
filters <- listGenomeFilters("hg19")

}
\seealso{
\code{germline_filters}
}
