% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGenomeFilters.R
\name{listRearFilters}
\alias{listRearFilters}
\title{List germline rearrangement filters derived from 10 lymphoblast
cell lines (mixed ethnicities) and 8 normal blood sammples.}
\usage{
listRearFilters(genome = c("hg19", "hg18"))
}
\arguments{
\item{genome}{length-one character vector indicating genome build (hg18 or hg19)}
}
\value{
a \code{GRangesList}
}
\description{
List germline rearrangement filters derived from 10 lymphoblast
cell lines (mixed ethnicities) and 8 normal blood sammples.
}
\examples{
listRearFilters("hg19")
}
\seealso{
\code{\link{reduceRearFilters}}
}
