% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{loadGenomeData}
\alias{loadGenomeData}
\title{Puts txdb, BSgenome, transcripts, and CDS in global environment}
\usage{
loadGenomeData(build = "hg19")
}
\arguments{
\item{build}{genome build -- only 'hg19' and 'hg18' supported}
}
\description{
Puts txdb, BSgenome, transcripts, and CDS in global environment
}
\examples{
rm(list=ls())
loadGenomeData()
exists(c("cds", "txdb", "tx", "genome"))
}
