% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{loadTx}
\alias{loadTx}
\title{Load Txdb object}
\usage{
loadTx(build)
}
\arguments{
\item{build}{genome build (one of hg18, hg19, or hg38)}
}
\value{
a \code{TxDb} object
}
\description{
Load Txdb object
}
\examples{
library(svfilters.hg19)
tx <- loadTx("hg19")
tx
}
