% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{loadTxDb}
\alias{loadTxDb}
\title{Loads TxDb object from TxDb.Hsapiens.UCSC.<build>.refGene}
\usage{
loadTxDb(build)
}
\value{
A \code{TxDb} object
}
\description{
Loads TxDb object from TxDb.Hsapiens.UCSC.<build>.refGene
}
\examples{
txdb <- loadTxDb("hg19")
txdb
}
