% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{loadTxdbTranscripts}
\alias{loadTxdbTranscripts}
\title{Load transcripts, CDS, and a txdb object for a user-specified build}
\usage{
loadTxdbTranscripts(build, seq_levels)
}
\arguments{
\item{build}{character string providing genome build. Currently only hg19 and hg18 are supported.}

\item{seq_levels}{allowed seqnames}
}
\value{
a named list with elements `transcripts` (\code{GRanges} of transcripts), `cds` (\code{GRangesList} of CDS listed by transcript), and `txdb` (a \code{txdb} object).
}
\description{
Load transcripts, CDS, and a txdb object for a user-specified build
}
\examples{
tx.cds <- loadTxdbTranscripts("hg19")
tx.cds[["cds"]]
tx.cds[["transcripts"]]
tx.cds[["txdb"]]
}
