% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{makeAGraph}
\alias{makeAGraph}
\title{Initialize a graph object for amplicon analyses}
\usage{
makeAGraph(segs, af, params)
}
\arguments{
\item{segs}{a \code{GRanges} object from the segmentation of the normalized coverage}

\item{af}{a list of germline filters}

\item{params}{a list of parameters}
}
\description{
Initialize a graph object for amplicon analyses
}
\examples{
  library(svfilters.hg19)
  library(svbams)
  library(Rsamtools)
  data(germline_filters)
  ##
  ## read in some CNVs
  ##
  cv.extdata <- system.file("extdata", package="svbams")
  segs <- readRDS(file.path(cv.extdata, "cgov44t_segments.rds"))
  extdata <- system.file("extdata", package="svbams")
  bview <- BamViews(bamPaths=file.path(extdata, "cgov44t_revised.bam"))
  params <- ampliconParams()
  germline_filters[["germline_cnv"]] <- GRanges()
  germline_filters[["outliers"]] <- GRanges()
  makeAGraph(segs, germline_filters, params)
}
\seealso{
\code{\link{ampliconParams}}
}
