% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Rearrangement-class.R,
%   R/RearrangementList-class.R
\name{modalRearrangement}
\alias{modalRearrangement}
\alias{modalRearrangement<-}
\alias{modalRearrangement,Rearrangement-method}
\alias{modalRearrangement<-,Rearrangement-method}
\alias{modalRearrangement,Rearrangement,ANY-method}
\alias{modalRearrangement,RearrangementList-method}
\title{Accessor for the modal rearrangement of a linked tag cluster}
\usage{
modalRearrangement(object)

modalRearrangement(object) <- value

\S4method{modalRearrangement}{Rearrangement}(object)

\S4method{modalRearrangement}{Rearrangement}(object) <- value

\S4method{modalRearrangement}{RearrangementList}(object)
}
\arguments{
\item{object}{a \code{Rearrangement} object}

\item{value}{a character-string indicating the modal rearrangement type}
}
\value{
a character string
}
\description{
For a two clusters of improper reads that are linked by the pairing
information, we classify the type of rearrangement supported by
each improper read pair.  Once each read pair is typed, we store
the modal type in the \code{Rearrangement} object.
}
