% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{noPrematureStop}
\alias{noPrematureStop}
\title{Assess whether there are any premature stop codons}
\usage{
noPrematureStop(fusion)
}
\arguments{
\item{fusion}{a \code{List} created by \code{partitionAASequence}}
}
\value{
a \code{LogicalList}
}
\description{
For a fusion to be in-frame, we require that the only stop codon is at the 3-prime end of the fusion amino acid sequence.
}
\seealso{
\code{\link{partitionAASequence}}
}
