% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R
\name{numberImproper}
\alias{numberImproper}
\title{Extract number of improper reads for each element in a StructuralVariant}
\usage{
numberImproper(object)
}
\arguments{
\item{object}{a \code{StructuralVariant}}
}
\description{
Extract number of improper reads for each element in a StructuralVariant
}
\examples{
data(deletions)
numberImproper(deletions)
}
