% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrangement-class.R, R/RearrangementList-class.R
\docType{methods}
\name{numberLinkingRP}
\alias{numberLinkingRP}
\alias{numberLinkingRP,Rearrangement-method}
\alias{numberLinkingRP,RearrangementList-method}
\title{Calculate the number of read pairs that link two clusters of improper reads}
\usage{
numberLinkingRP(object)

\S4method{numberLinkingRP}{Rearrangement}(object)

\S4method{numberLinkingRP}{RearrangementList}(object)
}
\value{
a numeric vector
}
\description{
Clusters of improper read pairs can be linked by the mates of the
improper reads that belong to these clusters.  This function
computes the number of improperly paired reads that link two
clusters. The number of read pairs that would link two clusters in
a bona fide somatic rearrangement depends on the depth of
sequencing.  For 30x coverage, we require at least 5 linking read
pairs.
}
\keyword{internal}

