% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{numberOverlappingDeletions}
\alias{numberOverlappingDeletions}
\title{Tabulate the frequency a gene has a deletion}
\usage{
numberOverlappingDeletions(tx, grl, maxgap = 5000)
}
\arguments{
\item{tx}{a \code{transcript} object as provided by the \code{svfilters} package}

\item{grl}{a \code{GRangesList} of deletions}

\item{maxgap}{a length-one numeric vector indicating the amount of space
between a deletion and a transcript allowed to consider overlapping. This
argument is passed to \code{overlapsAny}.}
}
\value{
a \code{data.frame} of gene names with frequencies

a \code{GRanges} object. **The coordinates are of the reduced transcript and not the observed deletion.**
}
\description{
Multiple hit genes. Deletions do not necessarily have to overlap so long as
the two deletions both hit the same gene.
}
\details{
1. list transcripts by gene
2. reduce the deletion granges for each subject to avoid over-counting
overlapping hemizygous deletions as 2 hits
3. count overlaps
}
\seealso{
\code{\link[IRanges]{overlapsAny}}
}
