% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{organizeReads}
\alias{organizeReads}
\title{Organize improper read pairs supporting a rearrangement into a data.frame}
\usage{
organizeReads(rlist, split_reads)
}
\arguments{
\item{rlist}{a `RearrangementList`}

\item{split_reads}{a data.frame of split reads from BLAT. See `rearrangedReads`.}
}
\description{
Facilitates plotting reads supporting a rearrangement with ggplot.
}
\seealso{
\code{\link{rearrangedReads}}
}
