% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{organizeReferenceByFrame}
\alias{organizeReferenceByFrame}
\title{Organize reference amino acid sequences by frame}
\usage{
organizeReferenceByFrame(proteins)
}
\arguments{
\item{proteins}{an object created by \code{translateCDS}}
}
\value{
a \code{List} of the reference amino acid sequences for each of the three possible reading frames
}
\description{
Organize reference amino acid sequences by frame
}
\seealso{
\code{\link{translateCDS}}
}
