% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{partitionAASequence}
\alias{partitionAASequence}
\title{Partition the amino acid sequence of the fusion protein into the sequences derived from the 5-prime and 3-prime transcripts}
\usage{
partitionAASequence(proteins)
}
\arguments{
\item{proteins}{a \code{List} of amino acid sequences}
}
\description{
By partioning the amino acid sequence of the fusion into its 5-prime and 3-prime parts, we can compare the 5-prime partition to the amino acid sequence of the 5-prime reference protein and the 3-prime partition to the amino acid sequence of the 3-prime reference protein.
}
\seealso{
\code{\link{translateCDS}}
}
