% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{paths}
\alias{paths}
\alias{paths<-}
\alias{paths,PreprocessViews2-method}
\alias{paths<-,PreprocessViews2-method}
\title{Accessor for file paths}
\usage{
paths(object)

paths(object) <- value
}
\arguments{
\item{object}{a \code{PreprocessViews2} object}

\item{value}{a character-vector of file paths to intermediate data}
}
\description{
File path to intermediate data saved to disk
}
