% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrangement-class.R,
%   R/RearrangementList-class.R
\docType{methods}
\name{percentRearrangement}
\alias{percentRearrangement}
\alias{percentRearrangement<-}
\alias{percentRearrangement,Rearrangement-method}
\alias{percentRearrangement<-,Rearrangement-method}
\alias{percentRearrangement,Rearrangement,ANY-method}
\alias{percentRearrangement,RearrangementList-method}
\title{The fraction of improper read pairs supporting the modal rearrangement type}
\usage{
percentRearrangement(object)

percentRearrangement(object) <- value

\S4method{percentRearrangement}{Rearrangement}(object)

\S4method{percentRearrangement}{Rearrangement}(object) <- value

\S4method{percentRearrangement}{RearrangementList}(object)
}
\arguments{
\item{object}{a \code{Rearrangement} object}

\item{value}{a numeric vector with range [0,1] indicating the
fraction of improper read pairs that support the modal
rearrangement}
}
\description{
The fraction of improper read pairs supporting the modal rearrangement type
}
\seealso{
\code{\link{modalRearrangement}}
}
