% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess-class.R, R/seqinfo-utils.R
\docType{class}
\name{PreprocessViews2-class}
\alias{PreprocessViews2}
\alias{PreprocessViews2-class}
\alias{rowRanges,PreprocessViews2,ANY-method}
\alias{rowRanges,PreprocessViews2-method}
\alias{rowRanges<-,PreprocessViews2,ANY-method}
\alias{seqinfo,PreprocessViews2-method}
\alias{seqinfo<-,PreprocessViews2,seqinfo-method}
\alias{seqinfo<-,PreprocessViews2-method}
\alias{seqlevels,PreprocessViews2-method}
\title{A container for storing views of preprocessed data}
\usage{
PreprocessViews2(object)

\S4method{rowRanges}{PreprocessViews2,ANY}(x) <- value

\S4method{rowRanges}{PreprocessViews2}(x, ...)

\S4method{seqlevels}{PreprocessViews2}(x)

\S4method{seqinfo}{PreprocessViews2}(x)

\S4method{seqinfo}{PreprocessViews2}(x) <- value
}
\arguments{
\item{object}{can be \code{missing} or an existing \code{BamViews} object}

\item{x}{a \code{PreprocessViews2} object}

\item{value}{an object of class \code{seqinfo}}

\item{...}{ignored}
}
\value{
A \code{PreprocessViews2} object
}
\description{
This class directly extends the \code{BamViews} class and provides
views to intermediate data (saved to disk) from the analysis of one
or more bam files.

Constructor for PreprocessViews2
}
\section{Slots}{

\describe{
\item{\code{scale}}{a length-one numeric vector.  We scale numeric data by
the value of \code{scale}, round to the nearest integer, and then
save as an integer.  This slot is for internal use only.}
}}
\examples{
PreprocessViews2()
paths(PreprocessViews2())
pviews <- PreprocessViews2()
paths(pviews) <- character()
paths(pviews)

}
\seealso{
See \code{\link{paths}} for the character-vector of file
  paths to the intermediate data.
}

