% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess-class.R
\docType{methods}
\name{setAs}
\alias{coerce,PreprocessViews2,RangedSummarizedExperiment-method}
\alias{setAs}
\title{Coerce a \code{PreprocessViews2} object to a \code{RangedSummarizedExperiment}}
\arguments{
\item{from}{character string ('PreprocessViews2')}

\item{to}{character string  ('RangedSummarizedExperiment')}
}
\value{
a \code{RangedSummarizedExperiment}
}
\description{
This method pulls the assay data from disk through the views object
interface, and then creates a \code{SummarizedExperiment} object
with an assay named 'copy'.
}
\examples{
pviews <- PreprocessViews2()
as(pviews, "RangedSummarizedExperiment")

}

