% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{promoterCDS}
\alias{promoterCDS}
\title{Create the full CDS of transcripts for those in which only the promoter of the transcript is involved in a fusion}
\usage{
promoterCDS(coding_jxns, build)
}
\arguments{
\item{coding_jxns}{a \code{GRanges} representation of the sequence junctions that occur in transcripts, possibly intronic.}

\item{cds}{a \code{GRangesList} of all CDS created by \code{cdsBy}}
}
\value{
\code{GRangesList}
}
\description{
Rearrangements in which the 5-prime genomic region involves only the promoter are named promoter_TX_NAME, where TX_NAME is the name of a transcript. This function returns a GRangesList of CDS, including CDS with name promoter_TX_NAME.
}
\seealso{
\code{\link{codingJunctions}} \code{\link[GenomicFeatures]{cdsBy}}
}
