% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R
\docType{methods}
\name{proper}
\alias{proper}
\alias{proper<-}
\alias{proper,StructuralVariant-method}
\alias{proper<-,StructuralVariant,GAlignmentPairs-method}
\alias{proper,StructuralVariant,GAlignmentPairs-method}
\title{Get / set properly paired reads}
\usage{
proper(object)

proper(object) <- value

\S4method{proper}{StructuralVariant}(object)

\S4method{proper}{StructuralVariant,GAlignmentPairs}(object) <- value
}
\arguments{
\item{value}{a \code{GAlignmentPairs} object}
}
\value{
\code{GAlignmentPairs}
}
\description{
Get / set properly paired reads
}
\keyword{internal}
