% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{properReadPairs_bedops}
\alias{properReadPairs_bedops}
\title{Proper read pairs near variant (properReadPairs for CRAM)}
\usage{
properReadPairs_bedops(prp4del.ga, del.gr, dp)
}
\arguments{
\item{prp4del.ga}{}

\item{del.gr}{}

\item{dp}{}
}
\value{
A GAlignmentPairs object that are proper reads near the deletion region
}
\description{
Proper read pairs near variant (properReadPairs for CRAM)
}
