% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{properReadPairs_bedops2}
\alias{properReadPairs_bedops2}
\title{Second part of properReadPairs_bedops}
\usage{
properReadPairs_bedops2(ga)
}
\arguments{
\item{ga}{}
}
\value{
A set of proper read pairs near deletion region
}
\description{
Used after we used bedmap of a bed file and the GRanges from prp_makeRthinned
}
