% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess-class.R
\name{rdsId}
\alias{rdsId}
\title{Helper for creating filenames with .rds extension}
\usage{
rdsId(x)
}
\arguments{
\item{x}{a \code{BamViews}-derived object}
}
\description{
Intermediate files are stored in directories given by
\code{DataPaths}.  The names of the intermediate files are formed
by concatenating the \code{colnames} of the \code{BamViews}-derived
object with the extension \code{.rds}.
}
\examples{
  library(Rsamtools)
  extdir <- system.file("extdata", package="Rsamtools", mustWork=TRUE)
  bamfile <- list.files(extdir, pattern="ex1.bam$", full.names=TRUE)
  bview <- BamViews(bamPaths=bamfile)
  rdsId(bview)


}
