% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{readBlat}
\alias{readBlat}
\title{Read output from the command-line blat}
\usage{
readBlat(filename, skip = 5, col_names = FALSE, ...)
}
\arguments{
\item{filename}{character string providing full path to blat output}

\item{...}{additional arguments to \code{read_tsv}}
}
\value{
a \code{data.frame} of alignment records from blat
}
\description{
The command-line version of blat can be downloaded from sourceforge:
\url{http://sourceforge.net/projects/blat/files/}
}
\seealso{
See \code{\link{blatScores}} for evaluating whether the
  blat alignments support a novel sequence junction
}
