% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAlignmentPairs-utils.R
\name{readPairsAsSegments}
\alias{readPairsAsSegments}
\title{Represent read pairs as segments}
\usage{
readPairsAsSegments(object)
}
\arguments{
\item{object}{a \code{GAlignmentPairs} object}
}
\value{
a \code{GRanges} object
}
\description{
When both reads of a pair are aligned to the same chromosome, this
function makes a \code{GRanges} object of the interval (DNA
fragment).
}

