% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{readPairsNearVariant}
\alias{readPairsNearVariant}
\title{Extract all mapped read pairs near a deletion(s)}
\usage{
readPairsNearVariant(object, bam.file)
}
\arguments{
\item{object}{a \code{GRanges} object}

\item{bam.file}{a character string providing complete path to bam file}
}
\value{
a \code{GAlignmentPairs} object
}
\description{
Wrapper for \code{readGAlignments} that extracts mapped read pairs
near deletions.
}
