% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rear_cds-data}
\alias{rear_cds-data}
\alias{rear_cds}
\title{An example Transcripts object}
\format{a \code{Transcripts} object}
\usage{
data(rear_cds)
}
\description{
An example Transcripts object
}
\examples{
data(rear_cds)
}
\keyword{datasets}
