% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rear_cgov7t}
\alias{rear_cgov7t}
\title{A RearrangementList from ovarian cell line CGOV7T}
\format{
a \code{RearrangementList} object
}
\usage{
data(rear_cgov7t)
}
\description{
A RearrangementList from ovarian cell line CGOV7T
}
\examples{
data(rear_cgov7t)
}
\keyword{datasets}
