% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rear_list-data}
\alias{rear_list-data}
\alias{rear_list}
\title{An example RearrangementList}
\format{a \code{RearrangementList}}
\usage{
data(rear_list)
}
\description{
A \code{RearrangementList} object.
}
\examples{
data(rear_list)
head(modalRearrangement(rear_list))
linkedBins(rear_list)
}
\keyword{datasets}
