% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{rearDataFrameList}
\alias{rearDataFrameList}
\title{Creates a data.frame with both possible 5 to 3 prime orientations}
\usage{
rearDataFrameList(rlist, build, maxgap = 5000)
}
\arguments{
\item{rlist}{a length-two list of orientations. Each element is an object of class \code{Rearrangement}}

\item{build}{character string -- only 'hg19' or 'hg18' supported}

\item{maxgap}{integer. See \code{findOverlaps}}
}
\description{
Creates a data.frame with both possible 5 to 3 prime orientations
}
