% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{rearrangedReadList}
\alias{rearrangedReadList}
\title{Identify rearranged reads -- initiallly unmapped reads that can be
aligned by blat to span a novel sequence junction.}
\usage{
rearrangedReadList(dirs, rlist, maxgap = 500)
}
\arguments{
\item{dirs}{a \code{DataPaths} object}

\item{rlist}{a list of \code{RearrangementList} objects}

\item{maxgap}{a length-one integer vector specifying how much gap
is allowed between a read alignment location and the genomic
intervals for the rearrangement clusters}
}
\value{
a list of rearranged reads
}
\description{
For unmapped-mapped read pairs where the mapped read is aligned
near a candidate junction, assess whether the unmapped read spans a
novel sequence junctions.
}
