% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{rearrangedReads}
\alias{rearrangedReads}
\title{Identify rearranged reads -- initiallly unmapped reads that can be
aligned by blat to span a novel sequence junction.}
\usage{
rearrangedReads(linked_bins, blat, maxgap = 500)
}
\arguments{
\item{linked_bins}{a \code{GRanges} of linked bins (e.g., gotten by \code{linkedBins(rearrangement.list)})}

\item{blat}{a data.frame of blat alignment records}

\item{maxgap}{this maximum gap between the mapped read and the
genomic intervals of the improper read clusters}
}
\value{
a `GRangesList` of blat records that map to both sides of a
sequence  junction. Each list element corresponds to one read that is
 aligned to two  locations (i.e., each element of the list consists of the
 vector of reads that supports one rearrangement).
}
\description{
Identify rearranged reads -- initiallly unmapped reads that can be
aligned by blat to span a novel sequence junction.
}
