% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Rearrangement-class.R,
%   R/rearrangement-utils.R
\docType{class}
\name{partitioning}
\alias{partitioning}
\alias{type}
\alias{Rearrangement-class}
\alias{Rearrangement}
\alias{partitioning,Rearrangement-method}
\alias{length,Rearrangement-method}
\alias{names,Rearrangement-method}
\alias{[,Rearrangement,ANY,ANY,ANY-method}
\alias{type,Rearrangement-method}
\alias{type,RearrangementList-method}
\title{Accessor for the combination of strands involved in a rearrangement}
\usage{
partitioning(object)

type(object)

\S4method{partitioning}{Rearrangement}(object)

\S4method{length}{Rearrangement}(x)

\S4method{names}{Rearrangement}(x)

\S4method{[}{Rearrangement,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{type}{Rearrangement}(object)

\S4method{type}{RearrangementList}(object)
}
\arguments{
\item{x}{a \code{Rearrangement} object}

\item{i}{a \code{logical} vector}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}
}
\value{
a \code{Rearrangement} object
}
\description{
This class contains the genomic intervals corresponding to improper
read clusters.  Improper read clusters that are linked are
candidate rearrangements.  The linking is represented by a
\code{GRanges} object with a 'linked.to' variable in the
\code{mcols}.  All improper reads belonging to a cluster,
irrespective of whether they link two clusters, are included and
indexed to facilitate subsetting operations.
}
\section{Slots}{

\describe{
\item{\code{linkedBins}}{a \code{GRanges} object with 'linked.to' in the
element metadata (\code{mcols})}

\item{\code{modal_rearrangement}}{character string indicating type of rearrangement}

\item{\code{improper}}{a \code{GAlignmentPairs} object of the improper read pairs}

\item{\code{partitioning}}{integer vector}

\item{\code{link1id}}{character string label for first read cluster}

\item{\code{link2id}}{character string label for second read cluster}

\item{\code{tags}}{a \code{GRanges} object of single tags}

\item{\code{tag_map_to_linked_bin}}{an integer vector keeping track of
which tags belong to a linked bin}

\item{\code{modal_rearrangement}}{charactering string of the modal
rearrangemnt type}

\item{\code{percent_rearrangement}}{length-one numeric vector indicating
the percentage of improper read pairs for a linked bin that
belong to the modal rearrangement}

\item{\code{fraction_linking_tags}}{the fraction of all tags for a linked
bin that link the two tag clusters}

\item{\code{split_reads}}{a \code{GRanges} object of the split read alignments}
}}

\keyword{internal}
