% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rearrangement_list}
\alias{rearrangement_list}
\title{An example RearrangementList object}
\format{a \code{RearrangementList} object}
\description{
A \code{RearrangementList} object contains genomic intervals that
are linked by improper read pairs where both reads were mapped.  In
addition, it also contains the improper reads.
}
\seealso{
\code{\link{RearrangementList}}
}
