% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{rearrangementData}
\alias{rearrangementData}
\title{Creates a list object of data required to identify rearrangements}
\usage{
rearrangementData(
  bins = GRanges(),
  read_pairs,
  rlist = RearrangementList(),
  filters = rFilters()
)
}
\arguments{
\item{bins}{a \code{GRanges} object of non-overlapping (typically 1kb) bins}

\item{read_pairs}{a \code{GAlignmentPairs}-representation of paired reads}

\item{rlist}{a \code{RearrangementList}}

\item{filters}{a \code{GRangesList} of sequence- and germline-based filters}
}
\value{
a list
}
\description{
Creates a list object of data required to identify rearrangements
}
