% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RearrangementList-class.R,
%   R/seqinfo-utils.R
\docType{class}
\name{RearrangementList}
\alias{RearrangementList}
\alias{RearrangementList-class}
\alias{RearrangementList,missing-method}
\alias{RearrangementList,list-method}
\alias{RearrangementList,Rearrangement-method}
\alias{colData,RearrangementList-method}
\alias{colData<-,RearrangementList,ANY-method}
\alias{colData,RearrangementList,ANY-method}
\alias{elementType,RearrangementList-method}
\alias{length,RearrangementList-method}
\alias{names,RearrangementList-method}
\alias{overlapsAny,RearrangementList,GRanges-method}
\alias{$<-,RearrangementList-method}
\alias{$,RearrangementList,ANY-method}
\alias{$,RearrangementList-method}
\alias{[,RearrangementList,ANY,ANY,ANY-method}
\alias{[[,RearrangementList,ANY,ANY-method}
\alias{[[<-,RearrangementList,ANY,ANY,ANY-method}
\alias{c,RearrangementList-method}
\alias{RearrangementList,}
\alias{c-method}
\alias{seqlevels,RearrangementList-method}
\title{Constructor for \code{RearrangementList} class}
\usage{
RearrangementList(object, modal_rearrangement, percent_rearrangement, colData)

\S4method{RearrangementList}{missing}(object, colData)

\S4method{RearrangementList}{list}(object, colData)

\S4method{RearrangementList}{Rearrangement}(object)

\S4method{colData}{RearrangementList}(x, ...)

\S4method{colData}{RearrangementList,ANY}(x) <- value

\S4method{elementType}{RearrangementList}(x, ...)

\S4method{length}{RearrangementList}(x)

\S4method{names}{RearrangementList}(x)

\S4method{overlapsAny}{RearrangementList,GRanges}(query, subject,
  maxgap = -1L, minoverlap = 0L, type = c("any", "start", "end", "within",
  "equal"), ...)

\S4method{$}{RearrangementList}(x, name) <- value

\S4method{$}{RearrangementList}(x, name)

\S4method{[}{RearrangementList,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{RearrangementList,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{RearrangementList,ANY,ANY,ANY}(x, i, j, ...,
  drop = FALSE) <- value

\S4method{c}{RearrangementList}(x, ..., recursive = FALSE)

\S4method{seqlevels}{RearrangementList}(x)
}
\arguments{
\item{object}{a list of \code{Rearrangement} objects}

\item{modal_rearrangement}{a character vector of the modal
rearrangement corresponding to each element in \code{object}.
Must be the same length as \code{object}.}

\item{percent_rearrangement}{a numeric vector indicating the
fraction of improper read pairs supporting the modal
rearrangement type.  Must be the same length as \code{object}.}

\item{colData}{a \code{DataFrame} containing metadata on the
\code{Rearrangement} objects.}

\item{x}{a \code{RearrangementList} object}

\item{...}{additional arguments ignored}

\item{query}{a \code{RearrangementList}}

\item{subject}{a \code{GRanges} object}

\item{i}{a numeric or character vector of rearrangement names}

\item{j}{ignored}

\item{drop}{ignored}

\item{...}{ignored}
}
\value{
a \code{RearrangementList} object

\code{RearrangementList}

a \code{Rearrangement} object

a \code{Rearrangement} object
}
\description{
Constructor for \code{RearrangementList} class

A given sample may have multiple rearrangements.  It is convenient
to encapsulate the data supporting each rearrangement in a
\code{Rearrangement} object, and to list all of these
rearrangements in a \code{RearrangementList}.
}
\section{Slots}{

\describe{
\item{\code{data}}{a list of \code{Rearrangement} objects}

\item{\code{elementType}}{a character vector}

\item{\code{names}}{a charcter vector of names.  Must have the same length as the list object.}

\item{\code{colData}}{a \code{DataFrame}}

\item{\code{modal_rearrangement}}{a \code{character} vector}

\item{\code{percent_rearrangement}}{a \code{numeric} vector}
}}

\examples{
RearrangementList()

}
\seealso{
\code{\linkS4class{Rearrangement}}

\code{\linkS4class{Rearrangement}}
}
\keyword{internal}
