% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/RearrangementParams-class.R,
%   R/RearrangmentParams-class.R
\docType{class}
\name{rpSeparation}
\alias{rpSeparation}
\alias{minNumberTagsPerCluster}
\alias{minClusterSize}
\alias{maxClusterSize}
\alias{minGapWidth}
\alias{percentModalType}
\alias{RearrangementParams-class}
\alias{RearrangementParams}
\alias{rpSeparation,RearrangementParams-method}
\alias{minNumberTagsPerCluster,RearrangementParams-method}
\alias{minClusterSize,RearrangementParams-method}
\alias{minNumberTagsPerCluster,RearrangementParams-method}
\alias{maxClusterSize,RearrangementParams-method}
\alias{minGapWidth,RearrangementParams-method}
\alias{percentModalType,RearrangementParams-method}
\alias{percentLinking}
\alias{RearrangementParams-class}
\alias{RearrangementParams}
\alias{rpSeparation}
\alias{minNumberTagsPerCluster}
\alias{minClusterSize}
\alias{maxClusterSize}
\alias{minGapWidth}
\alias{percentModalType}
\alias{rpSeparation,RearrangementParams-method}
\alias{minNumberTagsPerCluster,RearrangementParams-method}
\alias{minClusterSize,RearrangementParams-method}
\alias{minNumberTagsPerCluster,RearrangementParams-method}
\alias{maxClusterSize,RearrangementParams-method}
\alias{minGapWidth,RearrangementParams-method}
\alias{percentModalType,RearrangementParams-method}
\alias{percentLinking}
\title{Parameter class for rearrangement analysis}
\usage{
rpSeparation(object)

minNumberTagsPerCluster(object)

minClusterSize(object)

maxClusterSize(object)

minGapWidth(object)

percentModalType(object)

RearrangementParams(rp_separation = 10000, min_number_tags_per_cluster = 5,
  min_cluster_size = 115L, max_cluster_size = 5000L, min.gapwidth = 1000L,
  percent_modal_type = 0.9, percent_linking = 0.8)

\S4method{rpSeparation}{RearrangementParams}(object)

\S4method{minNumberTagsPerCluster}{RearrangementParams}(object)

\S4method{minClusterSize}{RearrangementParams}(object)

\S4method{maxClusterSize}{RearrangementParams}(object)

\S4method{minGapWidth}{RearrangementParams}(object)

\S4method{percentModalType}{RearrangementParams}(object)

percentLinking(object)

RearrangementParams(rp_separation = 10000, min_number_tags_per_cluster = 5,
  min_cluster_size = 115L, max_cluster_size = 5000L, min.gapwidth = 1000L,
  percent_modal_type = 0.9, percent_linking = 0.8)

rpSeparation(object)

minNumberTagsPerCluster(object)

minClusterSize(object)

maxClusterSize(object)

minGapWidth(object)

percentModalType(object)

\S4method{rpSeparation}{RearrangementParams}(object)

\S4method{minNumberTagsPerCluster}{RearrangementParams}(object)

\S4method{minClusterSize}{RearrangementParams}(object)

\S4method{maxClusterSize}{RearrangementParams}(object)

\S4method{minGapWidth}{RearrangementParams}(object)

\S4method{percentModalType}{RearrangementParams}(object)

percentLinking(object)
}
\arguments{
\item{object}{a \code{RearrangementParams} object}

\item{rp_separation}{length-one numeric vector indicating minimum
separation of the first and last read of a pair}

\item{min_number_tags_per_cluster}{length-one numeric vector
indicating the minimum number of reads in a cluster and the
minimum number of reads required to link two clusters}

\item{min_cluster_size}{length-one numeric vector; the minimum size
of a cluster of reads}

\item{max_cluster_size}{length-one numeric vector; the maximum size
of a cluster of reads}

\item{min.gapwidth}{length-one numeric vector; reads with at most
min.gapwidth separation between them are considered overlapping.}

\item{percent_modal_type}{length-one numeric vector; the percentage
of reads that must agree with the modal rearrangement type}

\item{percent_linking}{length-one numeric vector; two linked tag
clusters must be linked by at least this percentage of reads. See
details}

\item{rp_separation}{length-one numeric vector indicating minimum
separation of the first and last read of a pair}

\item{min_number_tags_per_cluster}{length-one numeric vector
indicating the minimum number of reads in a cluster and the
minimum number of reads required to link two clusters}

\item{min_cluster_size}{length-one numeric vector; the minimum size
of a cluster of reads}

\item{max_cluster_size}{length-one numeric vector; the maximum size
of a cluster of reads}

\item{min.gapwidth}{length-one numeric vector; reads with at most
min.gapwidth separation between them are considered overlapping.}

\item{percent_modal_type}{length-one numeric vector; the percentage
of reads that must agree with the modal rearrangement type}

\item{percent_linking}{length-one numeric vector; two linked tag
clusters must be linked by at least this percentage of reads. See
details}

\item{object}{a \code{RearrangementParams} object}
}
\description{
Parameter class for rearrangement analysis

Some details about why we look for linked tag clusters.

Parameter class for rearrangement analysis

Some details about why we look for linked tag clusters.
}
\details{
A tag cluster is defined as follows:

(i)   it must have at least <min_number_tags_per_cluster> reads

(ii) Provided (i) is TRUE, the cluster includes all improper reads
  with less than <min.gapwidth> separation

(iii) The size of a cluster is defined as the difference in the
  minimum basepair across all members and the maximum basepair
  across all members.  The size of the cluster must be at least
  <min_cluster_size> and no bigger than <max_cluster_size>.

Having determined the type of rearrangment supported by each read
  pair for two linked clusters, we require that the modal
  rearrangement type be supported by at least <percent_modal_type>>
  read pairs.

A tag cluster is defined as follows:

(i)   it must have at least <min_number_tags_per_cluster> reads

(ii) Provided (i) is TRUE, the cluster includes all improper reads
  with less than <min.gapwidth> separation

(iii) The size of a cluster is defined as the difference in the
  minimum basepair across all members and the maximum basepair
  across all members.  The size of the cluster must be at least
  <min_cluster_size> and no bigger than <max_cluster_size>.

Having determined the type of rearrangment supported by each read
  pair for two linked clusters, we require that the modal
  rearrangement type be supported by at least <percent_modal_type>>
  read pairs.
}
\section{Slots}{

\describe{
\item{\code{rp_separation}}{length-one numeric vector}

\item{\code{min_number_tags_per_cluster}}{length-one numeric vector}

\item{\code{min_cluster_size}}{length-one numeric vector}

\item{\code{max_cluster_size}}{length-one numeric vector}

\item{\code{min.gapwidth}}{length-one numeric vector}

\item{\code{percent_modal_type}}{length-one numeric vector}

\item{\code{percent_linking}}{length-one numeric vector}

\item{\code{rp_separation}}{length-one numeric vector}

\item{\code{min_number_tags_per_cluster}}{length-one numeric vector}

\item{\code{min_cluster_size}}{length-one numeric vector}

\item{\code{max_cluster_size}}{length-one numeric vector}

\item{\code{min.gapwidth}}{length-one numeric vector}

\item{\code{percent_modal_type}}{length-one numeric vector}

\item{\code{percent_linking}}{length-one numeric vector}
}}

\examples{

## Default rearrangement parameters for whole genome sequence data
##  with 30x coverage
rp <- RearrangementParams()


## Default rearrangement parameters for whole genome sequence data
##  with 30x coverage
rp <- RearrangementParams()

}
\keyword{internal}
