% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent.R
\name{recurrentDeletions}
\alias{recurrentDeletions}
\title{Aggregate deletions to the gene-level when evaluating recurrence}
\usage{
recurrentDeletions(tx, grl, maxgap = 5000)
}
\arguments{
\item{tx}{a \code{GRanges} object of transcripts}

\item{grl}{a \code{GRangesList} of deletions -- each element is a sample}

\item{maxgap}{length-one integer vector passed to \code{findOverlaps}}
}
\description{
Aggregate deletions to the gene-level when evaluating recurrence
}
\examples{
  genes <- GRanges("chr1", IRanges(5, 10))
  genes$gene_name <- "a"
  gr1 <- GRanges(rep("chr1", 2), IRanges(c(4, 8), c(6, 10)), id=rep("id1", 2))
  gr2 <- GRanges("chr1", IRanges(3, 9), id="id2")
  grl <- GRangesList(id1=gr1, id2=gr2)
  recurrentDeletions(genes, grl, maxgap=5)
}
