% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{recurrentDrivers}
\alias{recurrentDrivers}
\title{Table of recurrent drivers}
\usage{
recurrentDrivers(grl, transcripts, split = ", ")
}
\arguments{
\item{grl}{a \code{GRangesList} organized by sample}

\item{transcripts}{a \code{GRanges} object of transcripts as provided by the
\code{svfilters.<ucsc_build>} packages}

\item{split}{A character string indicating the string used to separate
drivers. By default, we assume the drivers associated with a given deletion
or amplicon are separated by a comma followed by a space.}
}
\value{
a \code{data.frame} of recurrent drivers
}
\description{
Given a GRangesList of amplicons or deletions (each element of list is a
sample), tabulate the frequency of driver deletions or amplifications.
}
