% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGenomeFilters.R
\name{reduceGenomeFilters}
\alias{reduceGenomeFilters}
\title{Create a reduced set of germline and sequence filters}
\usage{
reduceGenomeFilters(filters, seqlev)
}
\arguments{
\item{filters}{a list of germline CNV filters as GRanges objects}

\item{seqlev}{character vector of sequence names}
}
\value{
a reduced \code{GRanges} object of germline and sequence filters
}
\description{
Create a reduced set of germline and sequence filters
}
\details{
The reduced set is restricted to the sequence names provided in
  \code{seqlev} (if provided).
}
\examples{
library(svfilters.hg19)
filter.list <- listGenomeFilters("hg19")
filter.list <- filter.list[names(filter.list) != "transcripts"]
filters <- reduceGenomeFilters(filter.list)

}
\seealso{
\code{\link{listGenomeFilters}}

See \code{\link[GenomicRanges]{inter-range-methods}} for a
  description of \code{reduce}.
}
