% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGenomeFilters.R
\name{reduceRearFilters}
\alias{reduceRearFilters}
\title{Provides a reduced set of germline rearrangement filters derived
from 10 lymphoblast cell lines (mixed ethnicities) and 8 normal
blood sammples.}
\usage{
reduceRearFilters(filters, seqlev)
}
\arguments{
\item{filters}{a list of germline CNV filters}

\item{seqlev}{character vector of sequence names}
}
\value{
a \code{GRanges} object
}
\description{
TODO: deprecate
}
\details{
The reduced set is restricted to the sequence names provided in
  \code{seqlev} (if provided).
}
\examples{
filters <- listRearFilters("hg19")
reduceRearFilters(filters, "chr1")

}
\seealso{
\code{\link{listRearFilters}}
}
