% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{referenceProtein}
\alias{referenceProtein}
\title{Translate the unrearranged (reference) sequence}
\usage{
referenceProtein(genome, cds, nms)
}
\arguments{
\item{genome}{a \code{BSgenome} object}

\item{cds}{a \code{GRangesList} of transcript CDS}

\item{nms}{names of transcripts to translate}
}
\description{
Translate the unrearranged (reference) sequence
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg19)
genome <- BSgenome.Hsapiens.UCSC.hg19
data(rear_cds, package="trellis")
unrear_cds <- fullTranscripts(rear_cds)
##
## referenceProtein
##
ref.proteins <- referenceProtein(genome, unrear_cds, names(unrear_cds))
}
\seealso{
\code{\link{tumorProtein}}
}
