% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{removeAmbiguousAln}
\alias{removeAmbiguousAln}
\title{Removes rearrangements for which the BLAT-aligned reads do not pass QC}
\usage{
removeAmbiguousAln(rlist, blat_list)
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{blat_list}{a list of \code{data.frame} objects, as returned
by \code{scoreBlatExperiment}}
}
\description{
BLAT records for which the \code{passQC} variable is \code{FALSE}
are removed.  The \code{RearrangementList} is then subset to
include only those rearrangement ids that remain in the BLAT
\code{data.frame}.
}
\examples{
\dontrun{
 library(svovarian)
 dirs <- projectOvarian(rootname="OvarianData2")
 if(FALSE){
 ## A BLAT-filtered RearrangementList
 ##saved_result <- readRDS(file.path(dirs[["unit_test"]], "blat_filtered.rds"))
 ##tag_list <- readRDS(file.path(dirs[["unit_test"]], "tag_seqs.rds"))
 blat <- scoreBlatExperiment(tag_list, dirs)
 blat_list <- blat["CGOV2T"]

 rlist <- readRDS(file.path(dirs[["rear:filter"]], "CGOV2T.rds"))
 rlist <- list(CGOV2T=rlist)
 ## Another BLAT-filtered RearrangementList created by 'removeAmbigousAln'
 filtered_rlist <- removeAmbiguousAln(rlist, blat_list)
 print(filtered_rlist)
 }
}
}
\seealso{
See \code{blatScores} for the approach used to QC
  BLAT-aligned reads.
}
