% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{rpSupportedDeletions}
\alias{rpSupportedDeletions}
\title{Counts the number of improper read pairs supporting a deletion}
\usage{
rpSupportedDeletions(sv, param, bins)
}
\arguments{
\item{sv}{a \code{StructuralVariant} object}

\item{param}{a \code{DeletionParam} object}

\item{bins}{a \code{GRanges} object containing the bins used in preprocessing
with mcols value 'log_ratio'}
}
\value{
a character-vector of deletion calls
}
\description{
If the number of improper read pairs supporting a deletion exceeds
\code{minFlankingHemizgyous(param)}, a deletion is designated as
"hemizygous+" if hemizygous and "homozygous+" if homozygous.
}
