% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R, R/RearrangementList-class.R
\docType{methods}
\name{lapply,StructuralVariant-method}
\alias{lapply,StructuralVariant-method}
\alias{sapply,StructuralVariant-method}
\alias{sapply,RearrangementList-method}
\alias{lapply,RearrangementList-method}
\title{Functionals for structural variant classes}
\usage{
\S4method{lapply}{StructuralVariant}(X, FUN, ...)

\S4method{sapply}{StructuralVariant}(X, FUN, ..., simplify = TRUE,
  USE.NAMES = TRUE)

\S4method{sapply}{RearrangementList}(X, FUN, ..., simplify = TRUE,
  USE.NAMES = TRUE)

\S4method{lapply}{RearrangementList}(X, FUN, ...)
}
\arguments{
\item{X}{see \code{showMethods("sapply")}}

\item{FUN}{a function}

\item{...}{additional arguments to \code{FUN}}

\item{simplify}{logical}

\item{USE.NAMES}{logical}
}
\description{
Functionals for structural variant classes
}
\keyword{internal}
