% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{scale_x_mb}
\alias{scale_x_mb}
\title{Use MB scale in gg-style plots}
\usage{
scale_x_mb(...)
}
\arguments{
\item{...}{additional arguments to scale_x_continuous}
}
\description{
Use MB scale in gg-style plots
}
