% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blat-utils.R
\name{scoreBlatExperiment}
\alias{scoreBlatExperiment}
\title{Reads files output by the blat executable.}
\usage{
scoreBlatExperiment(id, blat, tags, dirs, thr = 0.8)
}
\arguments{
\item{id}{sample id}

\item{blat}{data.frame of blat records from the command-line blat tool}

\item{tags}{a \code{tbl_df} object of the read sequences}

\item{dirs}{a \code{DataPaths} object}

\item{thr}{a length-one numeric vector indicating the fraction of
improper reads that are of high quality for a given
rearrangement.}
}
\value{
a list of \code{tbl_df} objects
}
\description{
The blat executable is run with the default set of arguments using
a reference genome comprised of standard sequences without
alternates.  This function is a wrapper for \code{blatScores}. If a
file containing \code{blatScores} already exists, this function
only reads previously saved computations from disk.
}
\examples{
\dontrun{
 library(svovarian)
 dirs <- projectOvarian(rootname="OvarianData2")
 if(FALSE){
   tags <- readRDS(file.path(dirs[["3read"]], "CGOV2T.rds"))
   blat <- readBlat(file.path(dirs["1blat"], "CGOV2T.txt"))
   parsed <- scoreBlatExperiment(id, blat, tags, dirs)
   parsed
 }
}

}
\seealso{
See \code{blatScores}
}
