% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segmentBins}
\alias{segmentBins}
\title{Segment log2-transformed and GC-adjusted counts using Circular Binary Segmentation.}
\usage{
segmentBins(bins, param = SegmentParam(), ...)
}
\arguments{
\item{bins}{a \code{GRanges} object containing adjusted counts. Note, the
adjusted counts must be stored in a column named \code{log_ratio}.}

\item{param}{a \code{SegmentParam} object}

\item{...}{additional arguments to \code{segment} in the \code{DNAcopy} package}
}
\description{
Segmentation of the transformed counts is performed by the Circular
Binary Segmentation (CBS) algorithm implemented in the \code{DNAcopy}
package.
}
\details{
A default set of segmentation parameters are provided in 
\code{SegmentParam()} and can be altered by creating a new \code{SegmentParam} 
object if desired.  

Some users may wish to control additional parameters to the CBS algorithm and 
therefore \code{segmentBins} is designed to take any argument from \code{DNAcopy::segment}
as input (see examples).
}
\examples{
  library(svfilters.hg19)
  data(bins1kb)
  library(GenomeInfoDb)
  library(DNAcopy)
  bins1kb <- keepSeqlevels(bins1kb, "chr22", pruning.mode = "coarse")
  bins1kb$log_ratio <- c(rnorm(ceiling(length(bins1kb)/2), mean = 0, sd = 0.4), 
                        rnorm(floor(length(bins1kb)/2), mean = -1, sd = 0.4))
  segmentBins(bins1kb) # Using default segmentation parameters
  segmentBins(bins1kb, param = SegmentParam(alpha = 0.01, undo.splits = "sdundo", 
                                            undo.SD = 5, verbose = 1))
  segmentBins(bins1kb, param = SegmentParam(), 
              weights = abs(rnorm(length(bins1kb))))                                         
  
}
\seealso{
\code{\link[DNAcopy]{segment}} for description of circular binary
  segmentation and references therein; see
  \code{\link[svclasses]{SegmentParam-class}} for a description of the
  default parameters settings passed to the \code{segment} function. See
  \code{\link[svpreprocess]{binNormalize}} for obtaining normalized counts
  for segmentation.
}
