% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segmentExperiment}
\alias{segmentExperiment}
\title{Segments log2-transformed and GC-adjusted counts.}
\usage{
segmentExperiment(object, tree, param = SegmentParam(), ...)
}
\arguments{
\item{object}{A \code{PreprocessViews2} object}

\item{tree}{A directory tree for storing intermediate files.}

\item{param}{An object of class \code{SegmentParam}.}

\item{...}{Additional arguments are passed to the \code{segment}
function in the \code{DNAcopy} package.}
}
\value{
A \code{GRangesList} object.  Each element is the set of
  \code{GRanges} for a given sample.  Meta-columns of the
  \code{GRanges} elements are \code{seg.mean} (the segment mean)
  and \code{sample} (an id for the sample).  Note that
  \code{segmentExperiment} saves the \code{GRangesList} object as
  an intermediate file in the provided directory tree. Parameters
  to the \code{segment} function are passed by an instance of the
  \code{SegmentParam} class.
}
\description{
Segmentation of the transformed counts is performed by the circular
binary segmentation algorithm implemented in the \code{DNAcopy}
package.
}
\seealso{
\code{\link[DNAcopy]{segment}} for description of circular
  binary segmentation and references therein; see
  \code{\link{SegmentParam-class}} for a description of the default
  parameters settings passed to the \code{segment} function.
}
