% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentParam-class.R
\docType{class}
\name{SegmentParam-class}
\alias{SegmentParam-class}
\alias{SegmentParam}
\alias{cbs_alpha}
\alias{undo.splits}
\alias{undo.SD}
\alias{cbs_alpha,SegmentParam-method}
\alias{undo.splits,SegmentParam-method}
\alias{undo.SD,SegmentParam-method}
\title{A class for storing segmentation parameters}
\usage{
SegmentParam(alpha = 0.001, undo.splits = "sdundo", undo.SD = 2,
  verbose = 0)

cbs_alpha(x)

undo.splits(x)

undo.SD(x)

\S4method{cbs_alpha}{SegmentParam}(x)

\S4method{undo.splits}{SegmentParam}(x)

\S4method{undo.SD}{SegmentParam}(x)
}
\arguments{
\item{alpha}{a length-one numeric vector}

\item{undo.splits}{character string}

\item{undo.SD}{length-one numeric vector}

\item{verbose}{length-one numeric vector}

\item{x}{\code{SegmentParam} object}
}
\value{
\code{SegmentParam} object
}
\description{
A class for storing segmentation parameters

A parameters class for segmentation by DNAcopy
}
\section{Slots}{

\describe{
\item{\code{alpha}}{A length-one numeric vector}

\item{\code{undo.splits}}{A length-one character vector}

\item{\code{undo.SD}}{A length-one numeric vector}

\item{\code{verbose}}{A length-one numeric vector}
}}

\examples{
sp <- SegmentParam()
cbs_alpha(sp)
undo.splits(sp)
undo.SD(sp)

}
\seealso{
\code{\link[DNAcopy]{segment}}

A description for each of these argments can be found by reading
  the documentation for the \code{segment} function in the
  \code{DNAcopy} package.
}
