% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{seqJunctionNearTx}
\alias{seqJunctionNearTx}
\title{Evaluates whether sequence junction of a rearrangement is within 5000bp of a transcript}
\usage{
seqJunctionNearTx(rlist, build)
}
\arguments{
\item{rlist}{a \code{RearrangementList}}

\item{tx}{a \code{GRanges} object}
}
\value{
a logical vector of the same length as the \code{rlist} object
}
\description{
Only rearrangements where both sides of the sequence junction are near a transcript are evaluated as candidate in-frame fusions.
}
\examples{
loadGenomeData()
 extdata <- system.file("extdata", package="svrearrange")
 rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
 rlist <- readRDS(rfile)
 seqJunctionNearTx(rlist, tx)
}
