% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Preprocess-class.R
\docType{methods}
\name{setScale<-}
\alias{setScale<-}
\alias{getScale}
\alias{setScale<-,PreprocessViews2-method}
\alias{setScale,PreprocessViews2-method}
\title{Set numeric scale for \code{PreprocessViews2} object}
\usage{
setScale(x) <- value

getScale(x)

\S4method{setScale}{PreprocessViews2}(x) <- value
}
\arguments{
\item{x}{A \code{PreprocessViews2} object}

\item{value}{a length-one numeric vector}
}
\value{
numeric
}
\description{
Exported only for internal use by other packages.
}
\keyword{internal}
