% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleProcessSTM.R
\name{simpleProcessSTM}
\alias{simpleProcessSTM}
\alias{simpleProcessLDA}
\title{A simple script to train a STM base model and format for Trellis}
\usage{
simpleProcessSTM(corpus.path, k, out.path = NULL,
  ldavis.data.path = NULL)
}
\arguments{
\item{corpus.path}{Path to directory containing separate text files.}

\item{k}{Number of topics to train}

\item{out.path}{Path to output location for the Trellis model. If NULL, ignored.}

\item{ldavis.data.path}{Path to output location for term.frequency and doc.length. If NULL, ignored. See \code{trellis::toLDAvis}.}
}
\description{
TODO(tfs): Documentation. Is this even useful/good here?
}
\examples{
\dontrun{# TODO}
}
\seealso{
\pkg{\link{stm}}

\link[dest=https://ldavis.cpsievert.me/reviews/reviews.html]{ldavis}
}
