% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{sortByRead1}
\alias{sortByRead1}
\title{Sort GRanges object with read pairs R1 and R2 by start of R1}
\usage{
sortByRead1(gr)
}
\arguments{
\item{gr}{a \code{GRanges} object instantiated from a \code{GAlignmentPairs}}
}
\value{
a \code{GRanges} object
}
\description{
In order to plot the read pairs as position versus read pair index,
information on the mates needs to be kept intact and the visualization is
more clear if the reads are sorted by the first read in the pair. This
function turns the read tag in the 'id' field of the GRanges object to an
ordered factor. The levels of the factor are determined by the start position
of the first read (R1) in the pair.
}
\examples{
  library(svbams)
  library(TxDb.Hsapiens.UCSC.hg19.refGene)
  region <- GRanges("chr15", IRanges(63201003, 63209243))
  si <- seqinfo(TxDb.Hsapiens.UCSC.hg19.refGene)
  seqinfo(region) <- si["chr15", ]

  path <-system.file("extdata", package="svbams")
  bampath <- list.files(path, pattern="cgov44t_revised.bam$",
                        full.names=TRUE)

  iparams <- improperAlignmentParams(mapqFilter=30)
  pparams <- properAlignmentParams(mapqFilter=30)
  \dontrun{
    irp <- getImproperAlignmentPairs(bampath,
                                     iparams, build="hg19")
    g.irp <- ga2gr(irp, is.improper=TRUE)
    prp <- getProperAlignmentPairs(bampath,
                                   pparams, build="hg19")
    g.prp <- ga2gr(prp, is.improper=FALSE)
    gr <- c(g.irp, g.prp)
    gr <- sortByRead1(gr)
    gr
}
}
