% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrangement-class.R
\docType{methods}
\name{splitReads}
\alias{splitReads}
\alias{splitReads<-}
\alias{splitReads,Rearrangement-method}
\alias{splitReads<-,Rearrangement,GRanges-method}
\alias{splitReads,Rearrangement,GRanges-method}
\title{Accessor for split read alignments}
\usage{
splitReads(object)

splitReads(object) <- value

\S4method{splitReads}{Rearrangement}(object)

\S4method{splitReads}{Rearrangement,GRanges}(object) <- value
}
\arguments{
\item{object}{a \code{Rearrangement} object}

\item{value}{a \code{GRanges} object of split read alignments}
}
\description{
Accessor for split read alignments
}
