% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R, R/improper-methods.R, R/seqinfo-utils.R
\docType{class}
\name{readPairs}
\alias{StructuralVariant}
\alias{StructuralVariant-class}
\alias{[,StructuralVariant,ANY,ANY,ANY-method}
\alias{[,StructuralVariant,ANY,ANY-method}
\alias{[,StructuralVariant,numeric,ANY,ANY-method}
\alias{[,StructuralVariant,numeric,ANY-method}
\alias{[[,StructuralVariant,ANY,ANY-method}
\alias{improper,StructuralVariant-method}
\alias{length,StructuralVariant-method}
\alias{names,StructuralVariant-method}
\alias{readPairs}
\alias{readPairs,StructuralVariant-method}
\alias{seqinfo,StructuralVariant-method}
\alias{seqlengths,StructuralVariant-method}
\alias{sort,StructuralVariant-method}
\alias{variant,StructuralVariant-method}
\alias{variant<-,StructuralVariant,ANY-method}
\alias{variant<-,StructuralVariant-method}
\title{The StructuralVariant class stores data pertaining to somatic deletions}
\usage{
readPairs(object)

\S4method{[}{StructuralVariant,numeric,ANY,ANY}(x, i, j, ..., drop = FALSE)

StructuralVariant(variant = GRanges(), proper = .GAlignmentPairs(),
  improper = .GAlignmentPairs(), copynumber, calls, index_proper,
  index_improper, grouped_variant)

\S4method{length}{StructuralVariant}(x)

\S4method{names}{StructuralVariant}(x)

\S4method{readPairs}{StructuralVariant}(object)

\S4method{sort}{StructuralVariant}(x, decreasing = FALSE, ...)

\S4method{variant}{StructuralVariant}(object)

\S4method{variant}{StructuralVariant}(object) <- value

\S4method{[}{StructuralVariant,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{StructuralVariant,ANY,ANY}(x, i)

\S4method{improper}{StructuralVariant}(object)

\S4method{seqlengths}{StructuralVariant}(x)

\S4method{seqinfo}{StructuralVariant}(x)
}
\arguments{
\item{object}{a \code{StructuralVariant}}

\item{x}{a \code{StructuralVariant}}

\item{...}{ignored}

\item{variant}{a \code{GRanges} object of the deletion genomic intervals}

\item{proper}{a \code{GAlignmentPairs} object of proper read pairs near the deletion}

\item{improper}{a \code{GAlignmentPairs} object of improper read pairs near the deletion}

\item{copynumber}{the mean log2 ratio (derived from read depth) of the deletion region}

\item{calls}{character vector indicating the type of deletion
(hemizygous, hemizygous+, homozygous , homozygous+, overlapping
hemizygous)}

\item{index_proper}{integer vector used internally for subsetting}

\item{index_improper}{integer vector used internally for subsetting}

\item{grouped_variant}{a grouping factor for the deletion intervals}

\item{decreasing}{logical}
}
\description{
\code{StructuralVariant} is a vector-like class for storing genomic
intervals of somatic deletions and relevant supporting data.
Properly and improperly paired reads and indices so that subsetting
methods (\code{[}) extracts the relevant data for a given deletion.
}
\section{Slots}{

\describe{
\item{\code{variant}}{A  \code{GRanges} object}

\item{\code{proper}}{A  \code{GAlignmentPairs} object of properly paired reads}

\item{\code{improper}}{A \code{GAlignmentPairs} object of improperly paired
reads}

\item{\code{copynumber}}{A \code{numeric} vector of the mean log ratios
corresponding to the \code{variant} intervals. The numeric vector
must have the same length as \code{variant}.}

\item{\code{calls}}{A length-N \code{character} vector of deletion calls
where N is the number of intervals in \code{variant}.  Possible
calls are (i) hemizygous: hemizygous deletion with fewer than 5
supporting improper read pairs (ii) hemizygous+: hemizygous
deletion with 5 or more supporting improper read pairs (iii)
homozygous: homozygous deletion with fewer than 5 supporting
improper read pairs (iv) homozygous+: homozygous deletion with 5
or more supporting improper read pairs}

\item{\code{index_proper}}{a list of indices.  Each element is a vector of
integers.  The kth element are the elements of the
\code{GAlignmentPairs} object in slot \code{proper} supporting
the kth deletion.}

\item{\code{index_improper}}{A list of indices.  Each element is a vector
of integers.  The kth element are the elements of the
\code{GAlignmentPairs} object in slot \code{improper} supporting
the kth deletion.}

\item{\code{grouped_variant}}{a factor indicating whether deletions belong
to the same group (e.g., overlapping hemizygous deletions)}

\item{\code{length_improper}}{Keeps track of the numer of the total number
of improper read pairs for quick summaries of this object.}

\item{\code{length_proper}}{Keeps track of the numer of the total number of
proper read pairs for quick summaries of this object.}
}}
\examples{
StructuralVariant()


}
\keyword{internal}

