% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{sv_amp_makeQuery}
\alias{sv_amp_makeQuery}
\title{First part of sv_amplicons2_bedops}
\usage{
sv_amp_makeQuery(preprocess, amplicon_filters, params = ampliconParams())
}
\arguments{
\item{preprocess}{}

\item{amplicon_filters}{}

\item{params}{}
}
\value{
A set of GRanges near amplicons
}
\description{
First part of sv_amplicons2_bedops
}
