% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon-utils.R
\name{sv_amplicon_exp}
\alias{sv_amplicon_exp}
\title{Identify somatic amplicons and grouped amplicons}
\usage{
sv_amplicon_exp(dirs, bviews, grl, amplicon_filters,
  params = ampliconParams(), transcripts)
}
\arguments{
\item{dirs}{character-vector of file paths for storing intermediate
files}

\item{bviews}{A \code{BamViews} object}

\item{grl}{A \code{GRangesList} of the segmented genomes (each
element is the \code{GRanges} for a sample)}

\item{amplicon_filters}{A list of germline filters and parameters. See the \code{germline_filters} object in the package \code{svfilters.<build>}.}

\item{params}{a list of parameters for the amplicon analysis}

\item{transcripts}{a \code{GRanges} object of the transcripts.}
}
\description{
This function identifies focal, somatic amplifications.  As
amplicons are often grouped by the bridge-fusion-breakage cycles,
we represent the set of somatic amplicons identified for a sample
as a graph. The nodes of the graph are the amplicons and the edges
are the links between amplicons informed by paired reads. This
function is a wrapper for \code{sv_amplicons} that constructs an
\code{AmpliconGraph} for a given sample.  As with the other
*Experiment functions, \code{sv_amplicon_exp} saves the
\code{AmpliconGraph} computed for each sample as in intermediate
file for quick recall.  If the file exists, this function reads the
serialized R object from disk. To generate an \code{AmpliconGraph}
de novo, one must first delete the intermediate files (see
examples).
}
\seealso{
See \code{\linkS4class{AmpliconGraph}} for methods
  associated with the class and \code{\link{sv_amplicons}} for
  construction of an \code{AmpliconGraph} for a single sample.
}
