% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{sv_amplicons2_bedops}
\alias{sv_amplicons2_bedops}
\title{sv_amplicons2 for CRAM}
\usage{
sv_amplicons2_bedops(
  preprocess,
  amplicon_filters,
  params = ampliconParams(),
  bed_file = amp.ga
)
}
\arguments{
\item{preprocess}{}

\item{amplicon_filters}{}

\item{params}{}

\item{bed_file}{}
}
\value{
A set of amplicons
}
\description{
sv_amplicons2 for CRAM
}
