% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{sv_amplicons2_bedops2}
\alias{sv_amplicons2_bedops2}
\title{Second part of sv_amplicons2_bedops}
\usage{
sv_amplicons2_bedops2(
  preprocess,
  amplicon_filters,
  params = ampliconParams(),
  bed_file
)
}
\arguments{
\item{preprocess}{}

\item{amplicon_filters}{}

\item{params}{}

\item{bed_file}{}
}
\value{
A set of amplicons
}
\description{
Second part of sv_amplicons2_bedops
}
