% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{sv_del_makeQuery}
\alias{sv_del_makeQuery}
\title{First half of sv_deletion}
\usage{
sv_del_makeQuery(preprocess, gr_filters, param = DeletionParam())
}
\arguments{
\item{preprocess}{}

\item{gr_filters}{}

\item{param}{}
}
\value{
A GRanges object that can be used for bedmap in command-line programs
}
\description{
First half of sv_deletion
}
