% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{sv_deletions}
\alias{sv_deletions}
\title{Creates a StructuralVariant object}
\usage{
sv_deletions(preprocess, gr_filters, param = DeletionParam())
}
\arguments{
\item{preprocess}{a list of preprocessing summaries (see \code{preprocessData})}

\item{gr_filters}{a \code{GRanges} object of germline filters.  If the function is
called without specifying this argument then a default set of germline filters 
will be applied from \code{svfilters.hg19} if the \code{genome} slot in \code{preprocess}
is set to "hg19" or \code{svfilters.hg18} if \code{genome} is set to "hg18".}

\item{param}{a \code{DeletionParam} object}
}
\value{
a \code{StructuralVariant} object
}
\description{
Creates a \code{StructuralVariant} object encapsulating information
on deletions, including the genomic intervals, proper and improper
read pairs in the vicinity, and a classification for the type of
deletion based on the preprocessed estimates of read depth and the
improper read pair alignments.
}
\details{
Note: proper read pairs near a candidate deletion are randomly sampled to
reduce the size of the object. For reproducibility, set a seed prior to
running this function.
}
\examples{
data(pdata)
sv_deletions(pdata)
}
\seealso{
\code{\link{preprocessData}}
}
