% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{sv_deletions_bedops}
\alias{sv_deletions_bedops}
\title{Define new functions for sv_deletions_bedops}
\usage{
sv_deletions_bedops(preprocess, gr_filters, param = DeletionParam(), object)
}
\arguments{
\item{preprocess}{}

\item{gr_filters}{}

\item{param}{}

\item{object}{}
}
\value{
A set of candidate structural variants
}
\description{
Define new functions for sv_deletions_bedops
}
