% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baf-utils.R
\name{svAF}
\alias{svAF}
\title{Compute allele frequencies at germline heterozygous positions}
\usage{
svAF(normalBam, tumorBam, genome, positions, region, n = 50000,
  minCovNormal = 20, minCovTumor = 20, max_depth = 1000,
  minMafNormal = 0.3, minMafTumor = 0)
}
\arguments{
\item{normalBam}{The path to a bam file}

\item{tumorBam}{The path to a bam file}

\item{genome}{The genome build of \code{tumorBam} and \code{normalBam}.  Possible values are "hg38", "hg19", and "hg18".}

\item{positions}{A \code{GRanges} object consisting of the genomic regions of interest.  The default set of positions is 
the \code{snps} object from \code{svfilters.hg38}, \code{svfilters.hg19} or \code{svilters.hg18} depending on the user-specified \code{genome} argument.  The \code{snps}
object contains 1,000,000 positions that are frequently seen as heterozygous, therefore fewer positions are needed to find a sufficient 
number of heterozygous positions.  The \code{snps} object contains more than enough positions for tumor ploidy/ploidy analysis on well-covered
WGS bam files, although for bam files genertated from targeted capture-based sequencing data it is recommended to use all or most the \code{dbsnp150_snps}
object in \code{svfilters.hg38}, \code{svfilters.hg19} or \code{svfilters.hg18} to find a sufficient number of heterozygous positions as these objects contain
over 12,000,000 positions.}

\item{region}{If \code{region} is specified only SNPs in \code{position} that fall in \code{region} will 
be used.  This argument is useful for capture-based sequencing technologies (e.g. Whole-exome sequencing) where we tend to only have high enough
coverage to accurately call SNPs in the targeted regions.  By supplying a \code{GRanges} object containing the targeted 
regions this function avoids calculating coverage metrics and allele frequencies for low coverage off-target regions. Alternatively, 
users can provide the full path to a bed file in BED format.}

\item{n}{The number of positions to use for pileup in \code{normalBam}.  The \code{snps} objects 
in \code{svfilters.hg38}, \code{svfilters.hg19} and \code{svfilters.hg18} contain 1 million frequently heterozygous positions.  By specifying the \code{n} argument
a random sample of the positions object of length \code{n} will be used.  For tumor purity/ploidy
analysis 10,000 heterozygous positions well spread across the genome is typically plenty.  The 
default value of \code{n = 50000} is generally sufficient to achieve this on a 30X WGS bam file.  For sequencing data
from targeted capture protocols, it is recommended to use the full set of SNPs in \code{dbsnp150_snp} in \code{svfilters.hg38}, 
\code{svfilters.hg19} or \code{svfilters.hg18} to identify a suffient number of SNPs with high enough coverage.}

\item{minCovNormal}{The minimum coverage of a position in \code{normalBam} to be considered.}

\item{minCovTumor}{The minimum coverage of a position in \code{tumorBam} to be considered.}

\item{max_depth}{The maximum number of alignments considered for each position.}

\item{minMafNormal}{The minimum minor allele frequency (MAF) in order to consider a position 
as heterozygous in \code{normalBam}.}

\item{minMafTumor}{The minimum minor allele frequency (MAF) in order to consider a position 
as heterozygous in \code{tumorBam}.  It is recommended to set this value to at least 0.05 
when setting \code{normalBam = NULL} to avoid outputting homozygous positions.}
}
\value{
A data.frame with the following columns:

\strong{Chrom}: The chromosome of the event \cr
\strong{Pos}: The coordinate of the event \cr
\strong{RefBase}: The base in the reference genome (corresponds to the refUCSC column in dbSNP build 150) \cr
\strong{AltBase}: The other observed base \cr
\strong{Normal.Mut.Count}: The coverage of AltBase in normalBam \cr
\strong{Normal.Coverage}: The distinct coverage of RefBase + AltBase in normalBam \cr
\strong{Tumor.Mut.Count}: The coverage of AltBase in tumorBam \cr
\strong{Tumor.Coverage}: The distinct coverage of RefBase + AltBase in tumorBam \cr
\strong{Tumor.MAF}: The minor allele frequency of the event in tumorBam
}
\description{
Given a set of genomic coordinates this function identifies heterozygous positions in a bam file 
given by \code{normalBam} and retuns the minor allele frequency at these positions in a bam file 
specified by \code{tumorBam}.
}
\details{
If using this function to generate allele frequencies in a tumor 
sample at germline heterozygous positions identified in a matched normal sample then
\code{tumorBam} should point to the bam file for the tumor sample and \code{normalBam} 
should point to the bam file for its matched normal.
}
\examples{
extdir <- system.file("extdata", package="svbams")
bam1 <- file.path(extdir, "cgov10t.bam")
bam2 <- file.path(extdir, "cgov44t_revised.bam")

data(snps, package = "svfilters.hg19")
snps <- keepSeqlevels(snps, c("chr3", "chr5"), pruning.mode = "coarse")

svAF(normalBam=bam1, 
     tumorBam=bam2, 
     genome="hg19", 
     positions = snps, 
     n = 1000, 
     minCovNormal = 10, 
     minCovTumor = 10, 
     max_depth = 1e3, 
     minMafNormal = 0.3, 
     minMafTumor = 0)

}
