% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{svalignments}
\alias{svalignments}
\alias{svalignments-package}
\title{Functions to extract improper read pairs from bam files}
\description{
Functions to extract improper read pairs from bam files
}
