% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{svclasses}
\alias{svclasses}
\alias{svclasses-package}
\title{Classes and methods for structural variation analyses}
\description{
svclass provides infrastructure for structural variant analyses of sequencing data.
}

