% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{svcnvs}
\alias{svcnvs}
\alias{svcnvs-package}
\title{Identify copy number variants from log2-transformed and GC-adjusted read counts.}
\description{
Segments bin-level estimates of copy number.  Use both read pairs
and read depth to identify focal deletions and amplicons. For
amplicons, we use read pairs to identify grouped amplicons.
}
