% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{svpreprocess}
\alias{svpreprocess}
\alias{svpreprocess-package}
\title{Preprocessing for whole genome structural variant analyses}
\description{
Preprocessing includes defining bin-level summaries of relative copy number,
modeling GC content, and excluding regions of the genome that are difficult
to interrogate for copy number alterations.
}
