% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rear_reads.R
\name{tag_sequenceExperiment}
\alias{tag_sequenceExperiment}
\title{Extracts read sequences}
\usage{
tag_sequenceExperiment(dp, rlist, aviews, MAX = 25L)
}
\arguments{
\item{dp}{a \code{DataPaths} object}

\item{rlist}{a \code{RearrangementList}}

\item{aviews}{an \code{AlignmentViews2} object}

\item{MAX}{the maximum number of read pairs to extract for a
rearrangement.  If the number of read pairs supporting a
rearrangement is greater than MAX, a random sample of MAX
supporting read pairs is returned.}
}
\description{
This function extracts the sequences of reads from a bam file and
saves the result as an interemediate file.  If the intermediate
file already exists, the read sequences are read from disk.
}
\details{
This function calls \code{getReadAlignmentPairs}, the function
  originally called by \code{writeImproperAlignments2}.  The reason
  we call this function a second time is that the read sequences
  were not saved in the initial query.  Saving the sequences in the
  original query would prevent having to access the bam file a
  second time, though increase the size of the improper read files.
  We could use.  Another improvement in efficiency might come from
  parsing the original large bam into a smaller bam containing
  improper read pairs in which both mates were mapped.
}
