% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rearrangement-class.R
\docType{methods}
\name{tags}
\alias{tags}
\alias{tags<-}
\alias{tags,Rearrangement-method}
\alias{tags<-,Rearrangement-method}
\alias{tags,Rearrangement,ANY-method}
\title{Accessor for all the improper reads}
\usage{
tags(object)

tags(object) <- value

\S4method{tags}{Rearrangement}(object)

\S4method{tags}{Rearrangement}(object) <- value
}
\arguments{
\item{object}{a \code{Rearrangement} object}

\item{value}{a \code{GRanges} object of read alignments}
}
\description{
Accessor for improper reads belonging to one of the linked
bins. While a given read can only belong to one cluster (the
clusters are non-overlapping), a single cluster can be linked to
several other clusters.  Consequently, a single read (ignoring its
mate) can belong to several linked clusters. For visualizing the
reads supporting a rearrangement, it may also be important to carry
forward tags that do not support the rearrangement.
}
\keyword{internal}
