% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{thinProperPairs}
\alias{thinProperPairs}
\title{Thin proper read pairs to reduce overplotting}
\usage{
thinProperPairs(gr, thin = 10)
}
\arguments{
\item{gr}{a \code{GRanges} object instantiated from a \code{GAlignmentPairs} object}

\item{thin}{integer indicating how much to thin the properly paired reads.}
}
\value{
A \code{GRanges} object
}
\description{
This function provides a simple interface to
subsample a \code{GRanges} object of properly paired
reads to reduce overplotting.
}
\details{
Setting the parameter \code{thin} to 10 (default) will 
return a \code{GRanges} object with 1/10 the original number of 
properly paired reads in \code{gr}.
}
\examples{
library(svbams)
path <- system.file("extdata", package="svbams")
bam.file <- file.path(path, "cgov10t.bam")
prp <- getProperAlignmentPairs(bam.file)
pgr <- ga2gr(prp, is.improper=FALSE)
length(pgr)
pgr2 <- thinProperPairs(pgr, 10)
length(pgr2)
}
