% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletion-utils.R
\name{thinReadPairs}
\alias{thinReadPairs}
\title{Extract read pairs from a \code{StructuralVariant} object}
\usage{
thinReadPairs(object, max.out = 25000)
}
\arguments{
\item{object}{a \code{StructuralVariant} object}

\item{max.out}{length-one integer vector: the maximum number of read pairs to return}
}
\description{
This function extracts all read pairs from a StructuralVariant object and
then performs a filtering step. The filtering step thins the number of read
pairs with insert size less than 3kb.
}
