% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-utils.R
\name{threshold}
\alias{threshold}
\title{Threshold values of a numeric vector}
\usage{
threshold(x, lim = c(-Inf, Inf), amount = 0)
}
\arguments{
\item{x}{numeric vector}

\item{lim}{numeric vector of length 2 indicating the values at which to threshold \code{x}}

\item{amount}{scalar indicating how much to jitter the points at the threshold.}
}
\description{
Threshold the values of a numeric vector according to
user-specified limits.  Values exceeding the threshold are reset to
the threshold and jittered by an amount specified by the user to
reduce overplotting.
}
\examples{
x <- rnorm(10)
threshold(x, c(-0.5,0.5))
}
\seealso{
\code{\link{jitter}}
}
