% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPaths-class.R
\name{topics}
\alias{topics}
\title{List directories of intermediate data pipelines}
\usage{
topics(object)
}
\arguments{
\item{object}{a \code{DataPaths} object}
}
\value{
a \code{character}-vector
}
\description{
List directories of intermediate data pipelines
}
\examples{
dp <- DataPaths()
topics(dp)

}
\seealso{
\code{\link{DataPaths}}
}
