% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment-utils.R
\name{totalWidth}
\alias{totalWidth}
\title{Find total width of a \code{GRanges} object}
\usage{
totalWidth(object)
}
\arguments{
\item{object}{a \code{GRanges} object}
}
\value{
numeric
}
\description{
Adds the width of the intervals in a \code{GRanges} object.
}
\examples{
gr <- GRanges(c("chr1", "chr2"), IRanges(c(11, 11), c(1000, 1000)))
totalWidth(gr)
}
