% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Transcripts-class.R
\docType{methods}
\name{txA}
\alias{txA}
\alias{txB}
\alias{txC}
\alias{txD}
\alias{txB}
\alias{txC}
\alias{txD}
\alias{txA<-}
\alias{txB<-}
\alias{txC<-}
\alias{txD<-}
\alias{left}
\alias{right}
\alias{txA,Transcripts-method}
\alias{txB,Transcripts-method}
\alias{txC,Transcripts-method}
\alias{txD,Transcripts-method}
\alias{txA<-,Transcripts-method}
\alias{txA<-,Transcripts,ANY-method}
\alias{txB<-,Transcripts-method}
\alias{txB<-,Transcripts,ANY-method}
\alias{txC<-,Transcripts-method}
\alias{txC<-,Transcripts,ANY-method}
\alias{txD<-,Transcripts-method}
\alias{txD<-,Transcripts,ANY-method}
\title{Accessor for transcript from Transcripts-derived class}
\usage{
txA(object)

txB(object)

txC(object)

txD(object)

txA(object) <- value

txB(object) <- value

txC(object) <- value

txD(object) <- value

left(x, ...)

right(x, ...)

\S4method{txA}{Transcripts}(object)

\S4method{txB}{Transcripts}(object)

\S4method{txC}{Transcripts}(object)

\S4method{txD}{Transcripts}(object)

\S4method{txA}{Transcripts}(object) <- value

\S4method{txB}{Transcripts}(object) <- value

\S4method{txC}{Transcripts}(object) <- value

\S4method{txD}{Transcripts}(object) <- value
}
\arguments{
\item{object}{a \code{Transcripts}-derived object}

\item{value}{a \code{GRangesList}}

\item{x}{a \code{Transcripts}-derived object}

\item{...}{ignored}
}
\description{
Accessor for transcript from Transcripts-derived class
}
\examples{
data(rear_cds)
txA(rear_cds)
txB(rear_cds)
txC(rear_cds)
txD(rear_cds)
}
