% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Transcripts-class.R
\docType{class}
\name{identifier}
\alias{identifier}
\alias{Transcripts-class}
\alias{Transcripts}
\alias{identifier,Transcripts-method}
\alias{identifier<-,Transcripts-method}
\alias{identifier,Transcripts,ANY-methodsetMethod("identifier",}
\alias{"Transcripts",}
\alias{function(object)}
\alias{object@id)}
\alias{[[,Transcripts,ANY,ANY-method}
\alias{[[<-,Transcripts,ANY,ANY,GRangesList-method}
\title{Container for storing CDS near a sequence junction}
\usage{
identifier(object)

Transcripts(
  tA = GRangesList(),
  tB = GRangesList(),
  tC = GRangesList(),
  tD = GRangesList(),
  id = character()
)

\S4method{identifier}{Transcripts}(object)

\S4method{identifier}{Transcripts}(object) <- value

\S4method{[[}{Transcripts,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{Transcripts,ANY,ANY,GRangesList}(x, i, j) <- value
}
\arguments{
\item{x}{a \code{Transcripts} object}

\item{i}{integer vector}

\item{j}{integer vector}

\item{...}{ignored}

\item{drop}{ignored}
}
\description{
Container for storing CDS near a sequence junction
}
\section{Slots}{

\describe{
\item{\code{tA}}{a \code{GRangesList}}

\item{\code{tB}}{a \code{GRangesList}}

\item{\code{tC}}{a \code{GRangesList}}

\item{\code{tD}}{a \code{GRangesList}}

\item{\code{id}}{sample id}
}}

\examples{
Transcripts()
}
\keyword{-internal}
\keyword{internal}
