% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Transcripts-class.R
\docType{class}
\name{fusions}
\alias{fusions}
\alias{fusions<-}
\alias{fusionNames}
\alias{identifier<-}
\alias{numberFusions}
\alias{TranscriptsFusion-class}
\alias{TranscriptsFusion}
\alias{names,ClippedTranscripts-method}
\alias{names,ClippedTranscripts-class}
\alias{fusions,TranscriptsFusion-method}
\alias{fusions<-,TranscriptsFusion,list-method}
\alias{fusions,TranscriptsFusion,list-method}
\alias{fusionNames,TranscriptsFusion-method}
\alias{numberFusions,TranscriptsFusion-method}
\title{A container for fused transcripts}
\usage{
fusions(object)

fusions(object) <- value

fusionNames(object)

identifier(object) <- value

numberFusions(object)

TranscriptsFusion(object, fusions = list())

\S4method{names}{ClippedTranscripts}(x)

\S4method{fusions}{TranscriptsFusion}(object)

\S4method{fusions}{TranscriptsFusion,list}(object) <- value

\S4method{fusionNames}{TranscriptsFusion}(object)

\S4method{numberFusions}{TranscriptsFusion}(object)
}
\arguments{
\item{value}{}

\item{value}{character string}
}
\description{
Container for fused transcripts.
}
\section{Slots}{

\describe{
\item{\code{fusions}}{a \code{list} of fusions}
}}

\examples{
TranscriptsFusion()
}
\keyword{internal}
