% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{translateCDS}
\alias{translateCDS}
\title{Translate the CDS in each of the three possible reading frames}
\usage{
translateCDS(cds.list)
}
\arguments{
\item{cds.list}{a \code{List} of CDS for each fusion}
}
\description{
For each fusion, we translate the CDS in each of the three possible reading frames.
}
\seealso{
\code{\link{fuseCDS_Rlist}}
}
