% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{trellis}
\alias{trellis}
\alias{trellis-package}
\title{Classes and methods for structural variation analyses}
\description{
trellis provides infrastructure for structural variant analyses of sequencing data.
}
