% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{tumorProtein}
\alias{tumorProtein}
\title{Translate the sequence of a rearranged DNA sequence}
\usage{
tumorProtein(genome, fused.txlist)
}
\arguments{
\item{genome}{a \code{BSgenome} object}

\item{fused.txlist}{a \code{GRangesList}}
}
\description{
Translate the sequence of a rearranged DNA sequence
}
\examples{
data(rear_cds, package="trellis")
library(BSgenome.Hsapiens.UCSC.hg19)
genome <- BSgenome.Hsapiens.UCSC.hg19
fused_tx <- fuse(clip(rear_cds))
fused.proteins <- tumorProtein(genome, fused_tx)

}
\seealso{
\code{\link{referenceProtein}}
}
