% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{type_rearrangement}
\alias{type_rearrangement}
\title{Summary stats for rearrangement object}
\usage{
type_rearrangement(r)
}
\arguments{
\item{r}{a \code{Rearrangement} object}
}
\description{
Summary stats for rearrangement object
}
