% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges-utils.R
\name{uncouple}
\alias{uncouple}
\title{Uncouple linked bins}
\usage{
uncouple(linked_bins)
}
\arguments{
\item{linked_bins}{a \code{GRanges} representation of two genomic
intervals that are linked.}
}
\value{
a \code{GRanges} object
}
\description{
Two genomic intervals that are linked are represented by a
\code{GRanges} object where a single elemnt provides the genomic
interval of a bin and a genomic interval of a bin that it is linked
to.  The latter is included in the metadata (mcols) variable
'linked.to'.  This function represents every genomic interval as a
single element.  Hence, a \code{GRanges} of linked intervals with
length one would have length-two after uncoupling the bins.
}
\examples{
linked <- GRanges("chr1", IRanges(1,5))
linked$linked.to <- setNames(GRanges("chr1", IRanges(10, 20)), "b")
names(linked) <- "a"
uncouple(linked)

}
