% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprotFeatures}
\alias{uniprotFeatures}
\title{Extract features from uniprot database}
\usage{
uniprotFeatures(up, fusions, strwrap.width = 30)
}
\arguments{
\item{up}{uniprot features}

\item{fusions}{table of fusions with gene.5prime and gene.3prime columns}

\item{strwrap.width}{integer indicating how to wrap text for highly descriptive features in uniprot}
}
\description{
Extract features from uniprot database
}
\examples{
## See fusions vignette
}
