% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_CRAM.R
\name{unmapped_read_bedops}
\alias{unmapped_read_bedops}
\title{unmapped_read for CRAM}
\usage{
unmapped_read_bedops(query.bedops, maxgap = 500, path, sample_name)
}
\arguments{
\item{query.bedops}{}

\item{maxgap}{}

\item{path}{}

\item{sample}{}
}
\value{
Unmapped reads for BLAT
}
\description{
unmapped_read for CRAM
}
