% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{validFusions}
\alias{validFusions}
\title{Create a List of in-frame fusions with no premature stops}
\usage{
validFusions(fusion.list, cds.list, inframe.nostop, ref.frames)
}
\arguments{
\item{fusion.list}{a \code{List} created by \code{partitionAASequence}}

\item{cds.list}{a \code{List} created by \code{fuseCDS_Rlist}}

\item{inframe.nostop}{a \code{List} of \code{LogicalList}s.  True values indicate fusions that are both in-frame and have no premature stop codons.}

\item{ref.frames}{list of amino acid squences for reference proteins, organized by frame}
}
\description{
Create a List of in-frame fusions with no premature stops
}
\seealso{
\code{\link{partitionAASequence}} \code{\link{fuseCDS_Rlist}}
}
