% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletion-class.R
\name{variant}
\alias{variant}
\alias{variant<-}
\title{Extract deletion regions from a StructuralVariant object}
\usage{
variant(object)

variant(object) <- value
}
\arguments{
\item{object}{a \code{StructuralVariant}}

\item{value}{a \code{GRanges} object}
}
\value{
a \code{GRanges} object for the genomic intervals of the deletions
}
\description{
Extract deletion regions from a StructuralVariant object
}
