% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCorpusFiles.R
\name{writeCorpusFiles}
\alias{writeCorpusFiles}
\title{Given a lists/vectors of filenames, generate a directory of .txt files}
\usage{
writeCorpusFiles(fileContents, fileNames, outDirectory,
  forceDirectory = FALSE)
}
\arguments{
\item{fileContents}{List or vector of the text of all documents}

\item{fileNames}{List or vector, of the same size as fileContents (and in the same order)}

\item{outDirectory}{Directory in which to write .txt files}

\item{forceDirectory}{Logical flag. If TRUE, will create the output directory if it doesn't exist.}
}
\description{
Many packages and tutorials in R operate primarily on .CSV files
containing all text of a corpus. However, to avoid loading the entire
text corpus at once, Trellis requires a directory of individual text
files in order to read files within the main Trellis interface.
writeCorpusFiles takes a list or vector of files (and filenames)
and writes out each file individually as a .txt file.
}
\examples{
# Values from "sample_documents.RData":
# "filenames" and "filecontents"

writeCorpusFiles(filecontents, filenames, "./example_corpus", forceDirectory = TRUE)
}
