% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_functions_for_CRAMs.R
\name{add.info}
\alias{add.info}
\title{Generic function - add seqinfo to bed files}
\usage{
add.info(data)
}
\arguments{
\item{data}{a \code{GRanges} object}
}
\description{
Generic function - add seqinfo to bed files
}
\details{
bed files made from command-line programs such as Samtools does not contain sequence
informtion such as \code{seqlengths}, \code{isCircular} and \code{genome}. These information
are necessary for \code{segmentBins} function.
}
\examples{
  library(svfilters.hg19)
  data(bins1kb)
  library(GenomeInfoDb)
  library(DNAcopy)
  fc.bowtie <- data.table::fread(here("cgov44t_comparison", "bowtie", "cgov44t_revised_fragcount.bed"))
  colnames(fc.bowtie) <- c("chr", "start", "end", "gc", "map", "cnt")
  bins1kb.bowtie <- GenomicRanges::makeGRangesFromDataFrame(fc.bowtie, starts.in.df.are.0based = FALSE, keep.extra.columns = TRUE) #starts.in.df.are.0based = TRUE makes ranges start from 01 not 02
  bins.bedops.bowtie <- add.info(bins1kb.bowtie)

}
\seealso{
\code{\link[GenomicRanges]{GRanges}} for description information about 
the length and the names of the chromosomes.
}
