% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/DeletionList-class.R
\name{DeletionList}
\alias{DeletionList}
\alias{DeletionList,list-method}
\alias{DeletionList,GRangesList-method}
\alias{DeletionList,missing-method}
\title{GRangesList-derived class for deletions}
\usage{
DeletionList(object)

\S4method{DeletionList}{list}(object)

\S4method{DeletionList}{GRangesList}(object)

\S4method{DeletionList}{missing}(object)
}
\arguments{
\item{object}{see \code{showMethods("DeletionList")}}
}
\description{
Each element in the list are the set of deletions for a sample
}
