% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/New_functions_for_CRAMs.R
\name{dt2ga2}
\alias{dt2ga2}
\title{Data frame to GAlignment conversion 
this function prepares for conversion of GAlignment to GAlignmentPairs
makeGAlignmentPairs looks for: qname, rname, rnext, pos, pnext, flags}
\usage{
dt2ga2(dt)
}
\arguments{
\item{dt}{a \code{GRanges} object}
}
\description{
Data frame to GAlignment conversion 
this function prepares for conversion of GAlignment to GAlignmentPairs
makeGAlignmentPairs looks for: qname, rname, rnext, pos, pnext, flags
}
\details{
this function converts data frame into GAlignment object. This step is used
before converting GAlignment to GAlignmentPairs.
}
\examples{
  bed <- data.table::fread(here("cgov44t_comparison", "eland", "cgov44t_revised.bed"))
  colnames(bed) <- c("chr", "start", "end", "id", "score", "strand", "flag", "cigar", "mrnm", "mpos")
  ga <- dt2ga2(bed)

}
