% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{fiveTo3Prime}
\alias{fiveTo3Prime}
\title{Put the linked genomic intervals in 5-prime to 3-prime order with respect to the rearranged genome}
\usage{
fiveTo3Prime(r, build, maxgap = 5000)
}
\arguments{
\item{r}{a \code{Rearrangement} object}
}
\description{
Put the linked genomic intervals in 5-prime to 3-prime order with respect to the rearranged genome
}
\examples{
  extdata <- system.file("extdata", package="trellis")
  rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
  rlist <- readRDS(rfile)
  r <- rlist[[1]]
  r2 <- fiveTo3Prime(r)
  ggRearrange(r2[[1]])
}
\seealso{
\code{\link{rearDataFrame}} \code{\link{ggRearrange}}
}
