% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcripts-class.R
\docType{methods}
\name{fuse}
\alias{fuse}
\alias{fuse,ClippedTranscripts-method}
\alias{fuse,TranscriptsFusion-method}
\title{Join two clipped transcripts as a GRangesList object}
\usage{
fuse(object, nms)

\S4method{fuse}{TranscriptsFusion}(object, nms)

\S4method{fuse}{ClippedTranscripts}(object, nms)
}
\arguments{
\item{object}{a \code{Transcripts}-derived class}

\item{nms}{name of transcript (character-string)}
}
\value{
A \code{GRangesList}
}
\description{
If there are only two transcripts to join, the GRangesList object
has length one.  In general, the GRangesList object returned has
length R x C where R is the number of transcripts from the gene
5-prime of the fusion and C is the number of transcripts from the
gene 3' of the fusion.
}
\examples{
  data(rear_cds)
  clipped <- clip(rear_cds)
  ## The fused transcript contains exon ranks 2-9 of NM_001288715 and exon
  ## ranks 10-57 of NM_007118 
  fused <- fuse(clipped)
}
\seealso{
\code{link{clip}}
}

