% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion-utils.R
\name{fusionTable2}
\alias{fusionTable2}
\title{Create a table of in-frame fusions}
\usage{
fusionTable2(fusions)
}
\arguments{
\item{fusions}{a \code{List} created by \code{validFusions}}
}
\description{
Create a table of in-frame fusions
}
\seealso{
\code{\link{validFusions}}
}
