% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/alignment-utils.R
\docType{methods}
\name{getSequenceOfReads}
\alias{getSequenceOfReads}
\alias{getSequenceOfReads,Rearrangement-method}
\alias{getSequenceOfReads,RearrangementList-method}
\title{Extract sequences of reads supporting rearrangements from a bam file}
\usage{
getSequenceOfReads(object, bam.file, params = RearrangementParams(),
  MAX = 25L, build)

\S4method{getSequenceOfReads}{Rearrangement}(object, bam.file,
  params = RearrangementParams(), MAX = 25L, build)

\S4method{getSequenceOfReads}{RearrangementList}(object, bam.file,
  params = RearrangementParams(), MAX = 25L, build)
}
\arguments{
\item{object}{a \code{Rearrangement} or \code{RearrangementList} object}

\item{bam.file}{complete path to BAM file}

\item{params}{a \code{RearrangementParams} object}

\item{MAX}{the maximum number of read pairs to extract for a
rearrangement.  If the number of read pairs supporting a
rearrangement is greater than MAX, a random sample of MAX
supporting read pairs is returned.}

\item{build}{the reference genome buld that reads were aligned to.  Currently
supported builds include "hg19" and "hg18".}
}
\value{
a data.frame
}
\description{
Extract sequences of reads supporting rearrangements from a bam file
}
