% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{ggRearrange}
\alias{ggRearrange}
\title{ggplot wrapper for plotting reads supporting a rearrangement}
\usage{
ggRearrange(df, ylab = "Read pair index", basepairs = 400, num.ticks = 5)
}
\arguments{
\item{df}{a \code{data.frame} as created by \code{rearDataFrame}}

\item{ylab}{y-axis label}

\item{basepairs}{integer specifying size of window in basepairs}

\item{num.ticks}{integer specifying number of x-axis ticks}
}
\description{
ggplot wrapper for plotting reads supporting a rearrangement
}
\examples{
  data(rlist)
  extdata <- system.file("extdata", package="trellis")
  unmap.file <- file.path(extdata, "blat_unmapped.txt")
  blat_unmap <- readBlat(unmap.file)
  split_reads <- rearrangedReads(linkedBins(rlist),
                                 blat_unmap, 500)
  splitReads(rlist) <- split_reads
  rlist2 <- fiveTo3List(rlist, build="hg19")
  r <- rlist2[[1]]
  df <- rearDataFrame(r, "hg19")
  \dontrun{
    ggRearrange(df)
  }
}
\seealso{
\code{\link{rearDataFrame}}
}
