% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangement-utils.R
\name{ggRearrange}
\alias{ggRearrange}
\title{gg-style plot of reads supporting a rearrangement}
\usage{
ggRearrange(df)
}
\arguments{
\item{df}{a \code{data.frame} as created by \code{rearDataFrame}}
}
\description{
gg-style plot of reads supporting a rearrangement
}
\examples{
  extdata <- system.file("extdata", package="svrearrange")
  rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
  rlist <- readRDS(rfile)
  r <- rlist[[1]]
  df <- rearDataFrame(r, "hg19")
  ggRearrange(df)
}
\seealso{
\code{\link{rearDataFrame}}
}
