% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchTrellis.R
\name{launchTrellis}
\alias{launchTrellis}
\title{Launch the Trellis shiny app}
\usage{
launchTrellis()
}
\arguments{
\item{corpus.path}{Path to directory containing separate text files.}

\item{k}{Number of topics to train}

\item{out.path}{Path to output location for the Trellis model. If NULL, ignored.}

\item{ldavis.data.path}{Path to output location for term.frequency and doc.length. If NULL, ignored. See \code{trellis::toLDAvis}.}
}
\description{
TODO(tfs): documentation in R doc
}
\examples{
A simple script to train a STM base model and format for Trellis

TODO(tfs): Documentation. Is this even useful/good here?

\dontrun{
    TODO
}
}
\seealso{
\pkg{\link{stm}}

\link[dest=https://ldavis.cpsievert.me/reviews/reviews.html]{ldavis}
}
