% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rearlist-utils.R
\docType{methods}
\name{linkedTo}
\alias{linkedTo}
\alias{linkedTo,GRanges-method}
\alias{linkedTo,Rearrangement-method}
\alias{linkedTo,RearrangementList-method}
\title{A region linked by improperly paired reads}
\usage{
linkedTo(x)

\S4method{linkedTo}{GRanges}(x)

\S4method{linkedTo}{Rearrangement}(x)

\S4method{linkedTo}{RearrangementList}(x)
}
\arguments{
\item{x}{a \code{GRanges} object with value `linked.to` in the `mcols`}
}
\description{
A region linked by improperly paired reads
}
