% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_CRAM.R
\name{prp_makeRthinned}
\alias{prp_makeRthinned}
\title{A broken-up version of properReadPairs_bedops that can be used in linux system. This
is the first part where we extract a set of GRanges that are near deletion regions. 
The second half of the properReadPairs_bedops function is properReadPairs_bedops2}
\usage{
prp_makeRthinned(del.gr, dp)
}
\arguments{
\item{del.gr}{}

\item{dp}{}
}
\value{
GRanges that are near deletion regions
}
\description{
A broken-up version of properReadPairs_bedops that can be used in linux system. This
is the first part where we extract a set of GRanges that are near deletion regions. 
The second half of the properReadPairs_bedops function is properReadPairs_bedops2
}
