% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearlist-utils.R
\name{seqJunctions_Rlist}
\alias{seqJunctions_Rlist}
\title{Use split reads for each rearrangement to more precisely define sequence boundry}
\usage{
seqJunctions_Rlist(rlist)
}
\arguments{
\item{a}{\code{RearrangementList}}
}
\value{
a \code{GRanges} object of the 5-prime and 3-prime sequence junctions.  The 3-prime junctions are included in the '3p' field of the \code{GRanges} object
}
\description{
This function constructs a \code{GRanges} object of the 5-prime and 3-prime sequence junctions for a \code{RearrangementList} that has already been ordered by its two 5-prime to 3-prime orientation.
}
\examples{
 library(svrearrange)
 extdata <- system.file("extdata", package="svrearrange")
 rfile <- file.path(extdata, "CGOV11T_1.bam.rds")
 rlist <- readRDS(rfile)
 near.coding <- seqJunctionNearTx(rlist, tx)
 rlist2 <- fiveTo3List(rlist[near.coding], build="hg19")
 jxns <- seqJunctions_Rlist(rlist2)
 jxns
}
