% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Rearrangement-class.R,
%   R/Deletion-class.R, R/DeletionList-class.R, R/RearrangementList-class.R
\name{splitReads}
\alias{splitReads}
\alias{splitReads<-}
\alias{splitReads,Rearrangement-method}
\alias{splitReads<-,Rearrangement,GRanges-method}
\alias{splitReads,Rearrangement,GRanges-method}
\alias{splitReads,StructuralVariant-method}
\alias{splitReads<-,StructuralVariant,GRanges-method}
\alias{splitReads,StructuralVariant,GRanges-method}
\alias{splitReads<-,DeletionList,GRangesList-method}
\alias{splitReads,DeletionList,GRangesList-method}
\alias{splitReads,DeletionList-method}
\alias{splitReads<-,RearrangementList,GRangesList-method}
\alias{splitReads,RearrangementList,GRangesList-method}
\alias{splitReads,RearrangementList-method}
\title{Accessor for split read alignments}
\usage{
splitReads(object)

splitReads(object) <- value

\S4method{splitReads}{Rearrangement}(object)

\S4method{splitReads}{Rearrangement,GRanges}(object) <- value

\S4method{splitReads}{StructuralVariant}(object)

\S4method{splitReads}{StructuralVariant,GRanges}(object) <- value

\S4method{splitReads}{DeletionList,GRangesList}(object) <- value

\S4method{splitReads}{DeletionList}(object)

\S4method{splitReads}{RearrangementList,GRangesList}(object) <- value

\S4method{splitReads}{RearrangementList}(object)
}
\arguments{
\item{object}{a \code{Rearrangement} object or \code{StructuralVariant} object}

\item{value}{a \code{GRanges} object of split read alignments}
}
\description{
Accessor for split read alignments
}
