% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_for_CRAM.R
\name{sv_deletions_bedops2}
\alias{sv_deletions_bedops2}
\title{Second half of sv_deletions}
\usage{
sv_deletions_bedops2(preprocess, gr_filters, param = DeletionParam(), object)
}
\arguments{
\item{preprocess}{}

\item{gr_filters}{}

\item{param}{}

\item{object}{}
}
\value{
A set of candidate structural variants.
}
\description{
Second half of sv_deletions
}
