% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLDAvis.R
\name{toLDAvis}
\alias{toLDAvis}
\title{Convert flat Trellis model to LDAvis format}
\usage{
toLDAvis(model.file, doc.length, term.frequency, launch = FALSE)
}
\arguments{
\item{model.file}{.RData file containing Trellis model. Must include beta, theta, vocab. Leaves treated as flat model.}

\item{doc.length}{Number of tokens per document, as a vector.}

\item{term.frequency}{Term frequencies. Format TBD}

\item{launch}{Logical flag. If FALSE, toLDAvis returns the JSON created. Else, launches LDAvis}
}
\description{
TODO(tfs): documentation in R doc
}
\examples{
# TBD (how to deal with example data files?)
}
\seealso{
\pkg{\link{LDAvis}}

\pkg{\link{servr}}
}
