% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcripts-class.R
\docType{class}
\name{fusions}
\alias{fusions}
\alias{fusions<-}
\alias{fusionNames}
\alias{identifier<-}
\alias{numberFusions}
\alias{TranscriptsFusion-class}
\alias{TranscriptsFusion}
\alias{fusions,TranscriptsFusion-method}
\alias{fusions<-,TranscriptsFusion,list-method}
\alias{fusions,TranscriptsFusion,list-method}
\alias{fusionNames,TranscriptsFusion-method}
\alias{numberFusions,TranscriptsFusion-method}
\alias{[,TranscriptsFusion,ExonSubset,ANY,ANY-method}
\title{A container for fused transcripts}
\usage{
fusions(object)

fusions(object) <- value

fusionNames(object)

identifier(object) <- value

numberFusions(object)

TranscriptsFusion(object, fusions = list())

\S4method{fusions}{TranscriptsFusion}(object)

\S4method{fusions}{TranscriptsFusion,list}(object) <- value

\S4method{fusionNames}{TranscriptsFusion}(object)

\S4method{numberFusions}{TranscriptsFusion}(object)

\S4method{[}{TranscriptsFusion,ExonSubset,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{value}{}

\item{x}{a \code{TranscriptsFusion} object}

\item{i}{an integer-vector}

\item{j}{ignored}

\item{...}{ignored}

\item{drop}{ignored}

\item{value}{character string}
}
\value{
a \code{ClippedTranscripts} object
}
\description{
Container for fused transcripts.
}
\section{Slots}{

\describe{
\item{\code{fusions}}{a \code{list} of fusions}
}}

\examples{
TranscriptsFusion()
}
\keyword{internal}
